% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_11.R
\docType{data}
\name{Table10_11}
\alias{Table10_11}
\title{Table 10_11}
\format{
\itemize{
\item \strong{Y}
\item \strong{X2}
\item \strong{X3}
}
}
\usage{
data('Table10_11')
}
\description{
hypothetical data
}
\keyword{datasets}
