% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_12.R
\docType{data}
\name{Table10_12}
\alias{Table10_12}
\title{Table 10_12}
\format{
\itemize{
\item \strong{Y: }consumption
\item \strong{X2: }wage income
\item \strong{X3: }nonwage, nonfarm income
\item \strong{X4: }farm income
}
}
\usage{
data('Table10_12')
}
\description{
An Economic Model of the United States
Source: L. R. Klein and A. S. Goldberger, An Economic Model of the United States, 1929–1952, North Holland Publishing Company, Amsterdam, 1964, p. 131.
}
\keyword{datasets}
