% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_13.R
\docType{data}
\name{Table10_13}
\alias{Table10_13}
\title{Table 10_13}
\format{
\itemize{
\item \strong{Year}
\item \strong{CPI: }Consumer Price Index
\item \strong{GDP: }Gross domestic product
\item \strong{Imports}
}
}
\usage{
data('Table10_13')
}
\description{
U.S. Imports, GDP, and CPI, 1975–2005 (For all urban consumers; 1982–84 = 100, except as noted)
Source: Department of Labor, Bureau of Labor Statistics
}
\keyword{datasets}
