% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_14.R
\docType{data}
\name{Table10_14}
\alias{Table10_14}
\title{Table 10_14}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{Y: } new passenger cars sold (thousands), seasonally unadjusted
\item \strong{X2: }new cars, Consumer Price Index, 1967 = 100, seasonally unadjusted
\item \strong{X3: }Consumer Price Index, all items, all urban consumers, 1967 = 100, seasonally unadjusted
\item \strong{X4: }the personal disposable income (PDI), billions of dollars, unadjusted for seasonal variation
\item \strong{X5: }the interest rate, percent, finance company paper placed directly
\item \strong{X6: }the employed civilian labor force (thousands), unadjusted for seasonal variation.
}
}
\usage{
data('Table10_14')
}
\description{
Passenger Car Data
Source: Business Statistics, 1986, A Supplement to the Current Survey of Business, U.S. Department of Commerce.
}
\keyword{datasets}
