% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_15.R
\docType{data}
\name{Table10_15}
\alias{Table10_15}
\title{Table 10_15}
\format{
\itemize{
\item \strong{obs}
\item \strong{HRS: }average hours worked during the year
\item \strong{RATE: }average hourly wage (dollars)
\item \strong{ERSP: }average yearly earnings of spouse (dollars).
\item \strong{ERNO: } average yearly earnings of other family members (dollars).
\item \strong{NEIN: }average yearly nonearned income
\item \strong{ASSET: }average family asset holdings (bank account, etc.) (dollars)
\item \strong{AGE: }average age of responden
\item \strong{DEP: } average number of dependents
\item \strong{SCHOOL: }average highest grade of school completed.
}
}
\usage{
data('Table10_15')
}
\description{
Hours of Work and Other Data for 35 Groups
Source: D. H. Greenberg and M. Kosters, Income Guarantees and the Working Poor, Rand Corporation, R-579-OEO, December 1970
}
\keyword{datasets}
