% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_16.R
\docType{data}
\name{Table10_16}
\alias{Table10_16}
\title{Table 10_16}
\format{
\itemize{
\item \strong{Observation: }state (47 states for the year 1960)
\item \strong{R: }crime rate, number of offenses reported to police per million population
\item \strong{Age: }number of males of age 14–24 per 1,000 population.
\item \strong{S: }indicator variable for southern states (0 = no, 1 = yes)
\item \strong{ED: }mean number of years of schooling times 10 for persons age 25 or older
\item \strong{EX0: }1960 per capita expenditure on police by state and local government
\item \strong{EX1: }1959 per capita expenditure on police by state and local government.
\item \strong{LF: }labor force participation rate per 1,000 civilian urban males age 14–24
\item \strong{M: }number of males per 1,000 females.
\item \strong{N: }state population size in hundred thousands
\item \strong{NW: }number of nonwhites per 1,000 population.
\item \strong{U1: }unemployment rate of urban males per 1,000 of age 14–24
\item \strong{U2: }unemployment rate of urban males per 1,000 of age 35–39
\item \strong{W: }median value of transferable goods and assets or family income in tens of dollars.
\item \strong{X: }the number of families per 1,000 earnings 1⁄2 the median income
}
}
\usage{
data('Table10_16')
}
\description{
U.S. Crime Data for 47 States in 1960
Source: W. Vandaele, “Participation in Illegitimate Activities: Erlich Revisted,” in A. Blumstein, J. Cohen, and D. Nagin, eds., Deterrence and Incapacitation, National Academy of Sciences, 1978, pp. 270–335
}
\keyword{datasets}
