% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_17.R
\docType{data}
\name{Table10_17}
\alias{Table10_17}
\title{Table 10_17}
\format{
\itemize{
\item \strong{Observation}
\item \strong{Y: }number of people employed, in thousands
\item \strong{X1: }GNP implicit price deflator
\item \strong{X2: }GNP, millions of dollars
\item \strong{X3: }number of people unemployed in thousands
\item \strong{X4: }number of people in the armed forces in thousands
\item \strong{X5: }noninstitutionalized population over 16 years of age
\item \strong{X6: }year, equal to 1 in 1959, 2 in 1960, and 47 in 2005.
}
}
\usage{
data('Table10_17')
}
\description{
Updated Longley Data, 1959–2005
Source: Department of Labor, Bureau of Labor Statistics and http://siadapp.dmdc.osd.mil/personnel/MILITARY/Miltop.htm
}
\keyword{datasets}
