% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_18.R
\docType{data}
\name{Table10_18}
\alias{Table10_18}
\title{Table 10_18}
\format{
\itemize{
\item \strong{obs}
\item \strong{TASTE}
\item \strong{ACETIC: }the natural logarithm of acetic acid
\item \strong{H2S: } the natural logarithm of hydrogen sulfide
\item \strong{LACTIC}
}
}
\usage{
data('Table10_18')
}
\description{
Chemicals in Cheeses
Source: http://lib.stat.cmu.edu/DASL/Datafiles/Cheese.html
}
\keyword{datasets}
