% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_4.R
\docType{data}
\name{Table10_4}
\alias{Table10_4}
\title{Table 10_4}
\format{
\itemize{
\item \strong{Y}
\item \strong{X2}
\item \strong{X3}
}
}
\usage{
data('Table10_4')
}
\description{
Hypothetical Data on Y, X2, and X3
}
\keyword{datasets}
