% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_5.R
\docType{data}
\name{Table10_5}
\alias{Table10_5}
\title{Table 10_5}
\format{
\itemize{
\item \strong{Y}
\item \strong{X2} $
\item \strong{X3} $
}
}
\usage{
data('Table10_5')
}
\description{
Hypothetical Data on Consumption Expenditure Y, Income X2, and Wealth X3
}
\keyword{datasets}
