% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_7.R
\docType{data}
\name{Table10_7}
\alias{Table10_7}
\title{Table 10_7}
\format{
\itemize{
\item \strong{Year}
\item \strong{C: } consumption expenditure
\item \strong{Yd: }real disposable personal income
\item \strong{W: }real wealth
\item \strong{I: }real interest rate
}
}
\usage{
data('Table10_7')
}
\description{
U.S. Consumption Expenditure for the Period 1947–2000
Source: See table 7.12
}
\keyword{datasets}
