% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table10_8.R
\docType{data}
\name{Table10_8}
\alias{Table10_8}
\title{Table 10_8}
\format{
\itemize{
\item \strong{obs}
\item \strong{Y: }number of people employed, in thousands
\item \strong{X1: }GNP implicit price deflator
\item \strong{X2: }GNP, millions of dollars
\item \strong{X3: }number of people unemployed in thousands
\item \strong{X4: }number of people in the armed forces
\item \strong{X5: }noninstitutionalized population over 14 years of age
\item \strong{TIME} Year
}
}
\usage{
data('Table10_8')
}
\description{
Longley Data
Source: J. Longley, “An Appraisal of Least-Squares Programs from the Point of the User,” Journal of the American Statistical Association, vol. 62, 1967, pp. 819–841
}
\keyword{datasets}
