% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_1.R
\docType{data}
\name{Table11_1}
\alias{Table11_1}
\title{Table 11_1}
\format{
\itemize{
\item \strong{IND} Check the book in page 369 to see the categories of industries in this dataset
\item \strong{A: } 1-4
\item \strong{B: } 5-9
\item \strong{C: } 10-19
\item \strong{D: } 20-49
\item \strong{E: } 50-99
\item \strong{F: } 100-249
\item \strong{G: } 250-499
\item \strong{H: } 500-999
\item \strong{I: } 1.000-2.499
}
}
\usage{
data('Table11_1')
}
\description{
Compensation per Employee ($) in Nondurable Manufacturing Industries According to Employment Size of Establishment, 1958
Source: The Census of Manufacturers, U.S. Department of Commerce, 1958 (computed by author).
}
\keyword{datasets}
