% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_2.R
\docType{data}
\name{Table11_2}
\alias{Table11_2}
\title{Table 11_2}
\format{
\itemize{
\item \strong{Name: } Nome of Mutual Fund
\item \strong{Ei: } Average Annual Return %
\item \strong{Sigma: } Standart Deviation of Annual Return %
\item \strong{E_hat_i: } Obtained from the regression: Êi = 5.8194 + 0.4590 σi.
\item \strong{u_hat_i: }Absolute value of the residuals
\item \strong{Rank_u_hat: } The ranking is in ascending order of values
\item \strong{Rank_sigma: } The ranking is in ascending order of values
\item \strong{d: } Difference between Two Rankings
\item \strong{d2: } d^2
}
}
\usage{
data('Table11_2')
}
\description{
Rank Correlation Test of Heteroscedasticity
}
\keyword{datasets}
