% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_5.R
\docType{data}
\name{Table11_5}
\alias{Table11_5}
\title{Table 11_5}
\format{
\itemize{
\item \strong{obs} Each obsservation stands for a differente type of industry, see the book in page 397
\item \strong{SALES}
\item \strong{RD}
\item \strong{PROFITS}
}
}
\usage{
data('Table11_5')
}
\description{
Sales and Employment for Companies Performing Industrial R&D in the United States, by Industry, 2005 (values are in millions of dollars)
Source: National Science Foundation, Division of Science Resources Statistics, Survey of Industrial Research and Development: 2005 and the U.S. Census Bureau Annual Survey of Manufacturers, 2005.
}
\keyword{datasets}
