% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_6.R
\docType{data}
\name{Table11_6}
\alias{Table11_6}
\title{Table 11_6}
\format{
\itemize{
\item \strong{Year_Quart: }
\item \strong{X1_10: }
\item \strong{X10_25: }
\item \strong{X25_50: }
\item \strong{X50_100: }
\item \strong{X100_250: }
\item \strong{X250_1000: }
\item \strong{X1000_: }
}
}
\usage{
data('Table11_6')
}
\description{
sales/cash ratio in U.S. manufacturing industries classified by the asset size of the establishment for the period 1971–I to 1973–IV
Source: Quarterly Financial Report for Manufacturing Corporations, Federal Trade Commission and the Securities and Exchange Commission, U.S. government, various issues (computed).
}
\keyword{datasets}
