% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_7.R
\docType{data}
\name{Table11_7}
\alias{Table11_7}
\title{Table 11_7}
\format{
\itemize{
\item \strong{HP: }engine horsepower.
\item \strong{MPG: }average miles per gallon
\item \strong{VOL: }cubic feet of cab space
\item \strong{WT: } vehicle weight, hundreds of pounds
\item \strong{SP: } top speed, miles per hour.
}
}
\usage{
data('Table11_7')
}
\description{
Passenger Car Mileage Data
Source: U.S. Environmental Protection Agency, 1991, Report EPA/AA/CTAB/91-02
}
\note{
Observation = car observation number (Names of cars not disclosed)
}
\keyword{datasets}
