% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_8.R
\docType{data}
\name{Table11_8}
\alias{Table11_8}
\title{Table 11_8}
\format{
\itemize{
\item \strong{Years.in.Rank}
\item \strong{Year} Midpoint of the range
\item \strong{Year.2} Y^2
\item \strong{Count}
\item \strong{Median: } Median Salary
}
}
\usage{
data('Table11_8')
}
\description{
Median Salaries of Full Professors in Statistics, 2007
Source: American Statistical Association, “2007 Salary Report.
}
\keyword{datasets}
