% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table11_9.R
\docType{data}
\name{Table11_9}
\alias{Table11_9}
\title{Table 11_9}
\format{
\itemize{
\item \strong{COUNTRY}
\item \strong{Y: } Stock Prices - Rate of Change % per Year
\item \strong{X: } Consumer Prices - Rate of Change % per Year
}
}
\usage{
data('Table11_9')
}
\description{
Stock and Consumer Prices, Post–World War II Period (through 1969)
Source: Phillip Cagan, Common Stock Values and Inflation: The Historical Record of Many Countries, National Bureau of Economic Research, Suppl., March 1974, Table 1, p. 4.
}
\keyword{datasets}
