% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table12_10.R
\docType{data}
\name{Table12_10}
\alias{Table12_10}
\title{Table 12_10}
\format{
\itemize{
\item \strong{Observation}
\item \strong{RR: }real rate of return on common stocks at time
\item \strong{Growth: }output growth in period (t + 1)
\item \strong{Inflation: }inflation in period t(Inft)
}
}
\usage{
data('Table12_10')
}
\description{
Rate of Return, Output Growth and Inflation, United States, 1954–1981
}
\keyword{datasets}
