% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table12_4.R
\docType{data}
\name{Table12_4}
\alias{Table12_4}
\title{Table 12_4}
\format{
\itemize{
\item \strong{Y: }index of real compensation per hour, business sector (1992 = 100)
\item \strong{X: }index of output, business sector (1992 = 100)
}
}
\usage{
data('Table12_4')
}
\description{
Indexes of Real Compensation and Productivity, U.S., 1960–2005 (Index numbers,  1992 = 100;  quarterly data  seasonally adjusted)
Source: Economic Report of the President, 2007, Table B-49.
}
\keyword{datasets}
