% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table12_7.R
\docType{data}
\name{Table12_7}
\alias{Table12_7}
\title{Table 12_7}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{C: }12-month average U.S. domestic price of copper (cents per pound)
\item \strong{G: } annual gross national product ($, billions)
\item \strong{I: }12-month average index of industrial production
\item \strong{L: }12-month average London Metal Exchange price of copper (pounds sterling)
\item \strong{H: }number of housing starts per year (thousands of units)
\item \strong{A: }12-month average price of aluminum (cents per pound)
}
}
\usage{
data('Table12_7')
}
\description{
Determinants of U.S. Domestic Price of Copper, 1951–1980
Source: The data were collected by Gary R. Smith from sources such as American Metal Market, Metals Week, and U.S. Department of Commerce publications.
}
\keyword{datasets}
