% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table12_8.R
\docType{data}
\name{Table12_8}
\alias{Table12_8}
\title{Table 12_8}
\format{
\itemize{
\item \strong{Y: } Personal Consumption Expenditure, Billions of 1958 Dollars
\item \strong{X: } Time
\item \strong{YHAT: } YHAT_t = β0 + β1Xt + ut.
\item \strong{U: } Residuals
}
}
\usage{
data('Table12_8')
}
\description{
Personal Consumption Expenditure, Billions of 1958 Dollars
}
\keyword{datasets}
