% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table12_9.R
\docType{data}
\name{Table12_9}
\alias{Table12_9}
\title{Table 12_9}
\format{
\itemize{
\item \strong{Year}
\item \strong{Sales: }Annual data are averages of monthly, not seasonally adjusted, figures
\item \strong{Inventories: }Seasonally adjusted, end of period figures beginning 1982 are not comparable with earlier period
\item \strong{Ratio: } Invetories/Sales
}
}
\usage{
data('Table12_9')
}
\description{
Inventories and Sales in U.S. Manufacturing, 1950–1991 (millions of dollars)
Source: Economic Report of the President, 1993, Table B-53, p. 408
}
\keyword{datasets}
