% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table13_2.R
\docType{data}
\name{Table13_2}
\alias{Table13_2}
\title{Table 13_2}
\format{
\itemize{
\item \strong{Y.: }True Consumption Expenditure
\item \strong{X.: }True Income
\item \strong{Y: }Measured Consumption Expenditure
\item \strong{X: }Measured Income
\item \strong{E: } E = Y - Y.
\item \strong{W: } W = X - X.
\item \strong{U: } U = Y. - Y
}
}
\usage{
data('Table13_2')
}
\description{
Hypothetical Data
I am indebted to Kenneth J. White for constructing this example. See his Computer Handbook Using SHAZAM, for use with Damodar Gujarati, Basic Econometrics, September 1985, pp. 117–121.
}
\keyword{datasets}
