% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table13_3.R
\docType{data}
\name{Table13_3}
\alias{Table13_3}
\title{Table 13_3}
\format{
\itemize{
\item \strong{Year}
\item \strong{PPCE: }Per Capita Personal Consumption Expenditure
\item \strong{PDPI: }per Capita Personal Disposable Income
}
}
\usage{
data('Table13_3')
}
\description{
Per Capita Personal Consumption Expenditure (PPCE) and per Capita Personal Disposable Income (PDPI), U.S., 1970–2005
Source: Economic Report of the President, 2007
}
\keyword{datasets}
