% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table14_1.R
\docType{data}
\name{Table14_1}
\alias{Table14_1}
\title{Table 14_1}
\format{
\itemize{
\item \strong{N: }
\item \strong{Y: } Fee %
\item \strong{X: }Asset Value, billions of dollars
}
}
\usage{
data('Table14_1')
}
\description{
Advisory Fees Charged and Asset Size
}
\keyword{datasets}
