% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table14_2.R
\docType{data}
\name{Table14_2}
\alias{Table14_2}
\title{Table 14_2}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{POPULATION}
}
}
\usage{
data('Table14_2')
}
\description{
U.S. Population (Millions)
Source: Economic Report of the President, 2008
}
\keyword{datasets}
