% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table14_3.R
\docType{data}
\name{Table14_3}
\alias{Table14_3}
\title{Table 14_3}
\format{
\itemize{
\item \strong{Observation: } Year
\item \strong{GDP: }GDP is in millions of 1960 pesos
\item \strong{Labor: }Labor is in thousands of people
\item \strong{Capital: }Capital is in millions of 1960 pesos
}
}
\usage{
data('Table14_3')
}
\description{
Production Function Data for the Mexican Economy
Source: Victor J. Elias, Sources of Growth: A Study of Seven Latin American Economies, International Center for Economic Growth, ICS Press, San Francisco, 1992, Tables E-5, E-12, E-14.
}
\keyword{datasets}
