% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table14_4.R
\docType{data}
\name{Table14_4}
\alias{Table14_4}
\title{Table 14_4}
\format{
\itemize{
\item \strong{YSTAR}
\item \strong{X1}
\item \strong{X2}
}
}
\usage{
data('Table14_4')
}
\description{
Table 14_4
}
\keyword{datasets}
