% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_1.R
\docType{data}
\name{Table15_1}
\alias{Table15_1}
\title{Table 15_1}
\format{
\itemize{
\item \strong{FAMILY}
\item \strong{Y: }Y = 1 If owns home, 0 Otherwise
\item \strong{X: } Income X (Thousands of dollars)
}
}
\usage{
data('Table15_1')
}
\description{
Hypothetical Data on Home Ownership
}
\keyword{datasets}
