% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_10.R
\docType{data}
\name{Table15_10}
\alias{Table15_10}
\title{Table 15_10}
\format{
\itemize{
\item \strong{P}P are from Table 15.5;
\item \strong{I} I = F^-1(P) are estimated from the standard normal CDF
}
}
\usage{
data('Table15_10')
}
\description{
Estimating the Index Ii from the Standard Normal CDF
}
\keyword{datasets}
