% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_2.R
\docType{data}
\name{Table15_2}
\alias{Table15_2}
\title{Table 15_2}
\format{
\itemize{
\item \strong{Occupational_Group}
\item \strong{Smoking}
\item \strong{Mortality}
}
}
\usage{
data('Table15_2')
}
\description{
Actual Y, Estimated Y, and Weights wi for the Home Ownership Example
See the page 548 for more details about this dataset
}
\keyword{datasets}
