% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_22.R
\docType{data}
\name{Table15_22}
\alias{Table15_22}
\title{Table 15_22}
\format{
\itemize{
\item \strong{X: } Concentration, miligrams per liter
\item \strong{log..X.: } Log(x)
\item \strong{Ni: } Total
\item \strong{ni: } Death
\item \strong{p.hat.I: } P = n/N
}
}
\usage{
data('Table15_22')
}
\description{
Toxicity Study and Rotenone on Chrysanthemum Aphis
Source: D. J. Fennet, Probit Analysis, Cambridge University Press, London, 1964
}
\keyword{datasets}
