% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_23.R
\docType{data}
\name{Table15_23}
\alias{Table15_23}
\title{Table 15_23}
\format{
\itemize{
\item \strong{STUDENT}
\item \strong{Q: } GRE Aptitude Test Scores in Quantitative
\item \strong{V: } GRE Aptitude Test Scores in Verbal
\item \strong{ADM: } Admitted to Graduate Program (Yes = 1, No = 0)
}
}
\usage{
data('Table15_23')
}
\description{
GRE Scores
Source: Donald F. Morrison, Applied Linear Statistical Methods, Prentice-Hall, Inc., Englewood Cliffs, NJ, 1983, p. 279 (adapted)
}
\keyword{datasets}
