% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_26.R
\docType{data}
\name{Table15_26}
\alias{Table15_26}
\title{Table 15_26}
\format{
\itemize{
\item \strong{Y}
\item \strong{Y.}
\item \strong{X}
}
}
\usage{
data('Table15_26')
}
\description{
Hypothetical Data Set Generated by the Model Y = −1 +  3X + ε and Y* = 1 If Y > 0
Source: William E. Becker and Donald M. Waldman, “A Graphical Interpretation of Probit Coefficients,” Journal of Economic Education, Fall 1989, Table 1, p. 373
}
\keyword{datasets}
