% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_27.R
\docType{data}
\name{Table15_27}
\alias{Table15_27}
\title{Table 15_27}
\format{
\itemize{
\item \strong{c1 }
\item \strong{c2}
\item \strong{u}
\item \strong{v}
\item \strong{county}
\item \strong{age}
\item \strong{education}
\item \strong{married}
\item \strong{children}
\item \strong{select}
\item \strong{wagefull}
\item \strong{wage}
\item \strong{lw}
\item \strong{work}
\item \strong{lwf}
}
}
\usage{
data('Table15_27')
}
\description{
Table 15_27
}
\keyword{datasets}
