% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_28.R
\docType{data}
\name{Table15_28}
\alias{Table15_28}
\title{Table 15_28}
\format{
\itemize{
\item \strong{AGE}
\item \strong{AGEEDUC}
\item \strong{EDUC}
\item \strong{INCOME}
\item \strong{PCIGS79}
\item \strong{SMOKER}
\item \strong{EDU.INC}
}
}
\usage{
data('Table15_28')
}
\description{
Table 15_28
}
\keyword{datasets}
