% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_29.R
\docType{data}
\name{Table15_29}
\alias{Table15_29}
\title{Table 15_29}
\format{
\itemize{
\item \strong{Cancer}
\item \strong{Age}
\item \strong{HIGD}
\item \strong{Degre}
\item \strong{CHK}
\item \strong{AGPI}
\item \strong{MenAge}
\item \strong{Miscarriages}
\item \strong{Births}
\item \strong{Weight}
\item \strong{AgePeriod}
\item \strong{Marital.Status}
}
}
\usage{
data('Table15_29')
}
\description{
Table 15_29
}
\keyword{datasets}
