% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table15_4.R
\docType{data}
\name{Table15_4}
\alias{Table15_4}
\title{Table 15_4}
\format{
\itemize{
\item \strong{X: } Income in Thousands of dollars
\item \strong{N: } Number of Families at Income X
\item \strong{n: } Number of Families Owning a House
}
}
\usage{
data('Table15_4')
}
\description{
Hypothetical Data
Source: http://lib.stat.cmu.edu/ DASL/Datafiles/SmokingandCancer.html
}
\keyword{datasets}
