% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16_1.R
\docType{data}
\name{Table16_1}
\alias{Table16_1}
\title{Table 16_1}
\format{
\itemize{
\item \strong{I: }airline id
\item \strong{T: }year id
\item \strong{Q: }output, in revenue passenger miles, an index number
\item \strong{C: }total cost, in $1,000
\item \strong{PF: } fuel price
\item \strong{LF: }load factor, the average capacity utilization of the fleet
\item \strong{ln_C: }
\item \strong{ln_Q: }
\item \strong{ln_PF: }
\item \strong{ln_LF: }
}
}
\usage{
data('Table16_1')
}
\description{
costs of six airline firms for the period 1970–1984
William H. Greene, Econometric Analysis, 6th ed., 2008. Data are located at http://pages.stern.nyu.edu/~wgreen/Text/econometricanalysis.htm
}
\keyword{datasets}
