% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16_17.R
\docType{data}
\name{Table16_17}
\alias{Table16_17}
\title{Table 16_17}
\format{
\itemize{
\item \strong{COUNTRY}
\item \strong{Ctry_Code}
\item \strong{YEAR}
\item \strong{LGASPCAR}
\item \strong{LINCOMEP}
\item \strong{LRPMG}
\item \strong{LCARPCAP.}
}
}
\usage{
data('Table16_17')
}
\description{
Gasoline Demand in the OECD
B. H. Baltagi and J. M. Griffin, “Gasoline Demand in the OECD: An Application of Pooling and Testing Procedures,” European Economic Review, vol. 22, 1983, pp. 117–137. The data for 18 OECD countries for the years 1960–1978 can be obtained from: http://www.wiley.com/legacy/wileychi/baltagi/supp/Gasoline.dat, or from the textbook website, Table 16.17.
}
\keyword{datasets}
