% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table16_18.R
\docType{data}
\name{Table16_18}
\alias{Table16_18}
\title{Table 16_18}
\format{
\itemize{
\item \strong{Occupational_Group}
\item \strong{Smoking}
\item \strong{Mortality}
}
}
\usage{
data('Table16_18')
}
\description{
The Determinants of Aid in the Post-Cold War Era
Source: The article by Subhayu Bandyopadhyay and Howard J. Wall, “The Determinants of Aid in the Post-Cold War Era,” Review, Federal Reserve Bank of St. Louis, November/December 2007, vol. 89, number 6, pp. 533–547
}
\keyword{datasets}
