% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_10.R
\docType{data}
\name{Table17_10}
\alias{Table17_10}
\title{Table 17_10}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{Y: } Plant Expenditure
\item \strong{X2: } Sales
}
}
\usage{
data('Table17_10')
}
\description{
Source: Economic Report of the President, 1993. Data on Y from Table B-52, p. 407; data on X2 from Table 8-53, p. 408.
}
\keyword{datasets}
