% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_11.R
\docType{data}
\name{Table17_11}
\alias{Table17_11}
\title{Table 17_11}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{INVESTMENT}
\item \strong{SALES}
\item \strong{INTEREST}
}
}
\usage{
data('Table17_11')
}
\description{
Investments, Sales, and Interest Rate, United States, 1960–1999
Source: Economic Report of the President, 2001, Tables B-18, B-57, and B-73.
}
\keyword{datasets}
