% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_12.R
\docType{data}
\name{Table17_12}
\alias{Table17_12}
\title{Table 17_12}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{COMPENS: }index of real compensation per hour (1992 = 100)
\item \strong{PRODUC: }index of output per hour (1992 = 100).
\item \strong{UNRATE: }civilian unemployment rate, %
}
}
\usage{
data('Table17_12')
}
\description{
Compensation, Productivity and Unemployment Rate, United States, 1960–1999
Source: Economic Report of the President, 2001, Table B-49, p. 332
}
\keyword{datasets}
