% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_13.R
\docType{data}
\name{Table17_13}
\alias{Table17_13}
\title{Table 17_13}
\format{
\itemize{
\item \strong{Year}
\item \strong{PC: }
\item \strong{PDI: }
\item \strong{Grossinv: }
\item \strong{GNP: }
\item \strong{LTI: }
}
}
\usage{
data('Table17_13')
}
\description{
Macroeconomic Data for the Greek Economy, 1960–1995
Source: H. R. Seddighi, K. A. Lawler, and A. V. Katos, Econometrics: A Practical Approach, Routledge, London, 2000, p. 158
}
\keyword{datasets}
