% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_2.R
\docType{data}
\name{Table17_2}
\alias{Table17_2}
\title{Table 17_2}
\format{
\itemize{
\item \strong{Year}
\item \strong{PPCE: }per capita personal consumption expenditure in chained 2000 dollars.
\item \strong{PPDI: }per capita personal disposable income in chained 2000 dollars
}
}
\usage{
data('Table17_2')
}
\description{
PPCE and PPDI, 1959–2006
Source: Economic Report of the President, 2007, Table B-31
}
\keyword{datasets}
