% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_5.R
\docType{data}
\name{Table17_5}
\alias{Table17_5}
\title{Table 17_5}
\format{
\itemize{
\item \strong{OBS}
\item \strong{M1: }C$, millions
\item \strong{R: }90-day prime corporate interest rate, %
\item \strong{P: }implicit price deflator (1981 = 100)
\item \strong{GDP: } C$, millions (1981 prices)
}
}
\usage{
data('Table17_5')
}
\description{
Money, Interest Rate, Price Index, and GDP, Canada
Source: Rao, op. cit.,pp. 210–213
}
\keyword{datasets}
