% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_7.R
\docType{data}
\name{Table17_7}
\alias{Table17_7}
\title{Table 17_7}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{PCON: } private consumption expenditure.
\item \strong{GDP: } gross domestic product
}
}
\usage{
data('Table17_7')
}
\description{
Private Consumption Expenditure and GDP, Sri Lanka
Source: The data are obtained from the data disk in Chandan Mukherjee, Howard White, and Marc Wuyts, Econometrics and Data Analysis for Developing Countries, Routledge, New York, 1998. The original data are from World Bank’s World Tables
}
\keyword{datasets}
