% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table17_8.R
\docType{data}
\name{Table17_8}
\alias{Table17_8}
\title{Table 17_8}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{INVETORY}
\item \strong{SALES}
}
}
\usage{
data('Table17_8')
}
\description{
Smoking and Lung Cancer
Source: Economic Report of the President, 2001, Table B-57, p. 340. The Z ’s are as shown in Eq. (17.13.13).
}
\keyword{datasets}
