% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table18_2.R
\docType{data}
\name{Table18_2}
\alias{Table18_2}
\title{Table 18_2}
\format{
\itemize{
\item \strong{Observation}
\item \strong{C: }personal consumption expenditure
\item \strong{I: }gross private domestic investment
\item \strong{Y: }gross domestic product
}
}
\usage{
data('Table18_2')
}
\description{
Source: Economic Report of the President, 2008, Table B-2
}
\keyword{datasets}
