% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table18_3.R
\docType{data}
\name{Table18_3}
\alias{Table18_3}
\title{Table 18_3}
\format{
\itemize{
\item \strong{YR.QTR}
\item \strong{OBS}
\item \strong{CARSALES}
\item \strong{PERSINCOME}
\item \strong{PRICEGAS}
\item \strong{QUANTGAS}
}
}
\usage{
data('Table18_3')
}
\description{
Smoking and Lung Cancer
Source: http://lib.stat.cmu.edu/ DASL/Datafiles/SmokingandCancer.html
}
\keyword{datasets}
