% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table18_4.R
\docType{data}
\name{Table18_4}
\alias{Table18_4}
\title{Table 18_4}
\format{
\itemize{
\item \strong{Year: }date
\item \strong{qtr: }quarter
\item \strong{realgdp: }real GDP (billions of dollars)
\item \strong{realcons: }real consumption expenditure
\item \strong{realinvs: }real investment by private sector
\item \strong{realgovt: }real government expenditure
\item \strong{realdpi: }real disposable personal income
\item \strong{cpi_u: }consumer price index
\item \strong{M1: }nominal money stock
\item \strong{tbilrate: }quarterly average of month-end 90-day T-bill rate
\item \strong{unemp: } Unemployment rate
\item \strong{pop: }population, millions, interpolate of year-end figures using constant growth rate per quarter
\item \strong{infl: }rate of inflation (first observation is missing)
\item \strong{realint: }expost real interest rate (first observation missing)
}
}
\usage{
data('Table18_4')
}
\description{
Source: These data are originally from the Department of Commerce, Bureau of Economic Analysis, and from www.economagic.com, and are reproduced from William H. Greene, Econometric Analysis, 6th ed., 2008, Table F5.1, p.1083.
}
\keyword{datasets}
