% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_1.R
\docType{data}
\name{Table1_1}
\alias{Table1_1}
\title{Table 1.1
U.S. EGG PRODUCTION}
\format{
A data.frame with 50 observations on 5 variables
\itemize{
\item \strong{State}
\item \strong{Y1:} eggs produced in 1990 (millions)
\item \strong{Y2:} eggs produced in 1991 (millions)
\item \strong{X1:} price per dozen (cents) in 1990
\item \strong{X2:} price per dozen (cents) in 1991
}
}
\usage{
data('Table1_1')
}
\description{
Source: World Almanac, 1993, p. 119. The data are from the Economic Research Service, U.S. Department of Agriculture.
}
\keyword{datasets}
