% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_3.R
\docType{data}
\name{Table1_3}
\alias{Table1_3}
\title{Table 1.3
CPI in Seven Industrial Countries, 1980 – 2005 (1982–1984 = 100)}
\format{
\itemize{
\item \strong{Year}
\item \strong{USA}
\item \strong{Canada}
\item \strong{Japan}
\item \strong{France}
\item \strong{Germany}
\item \strong{Italy}
\item \strong{UK}
}
}
\usage{
data('Table1_3')
}
\description{
Source: Economic Report of the President, 2007, Table 108, p. 354
}
\keyword{datasets}
