% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_4.R
\docType{data}
\name{Table1_4}
\alias{Table1_4}
\title{Table 1.4
Exchange Rates for Nine Countries: 1985 – 2006}
\format{
\itemize{
\item \strong{Year}
\item \strong{Australia}
\item \strong{Canada}
\item \strong{China}
\item \strong{Japan}
\item \strong{Mexico}
\item \strong{South Korea}
\item \strong{Sweden}
\item \strong{Switzerland}
\item \strong{UK}
}
}
\usage{
data('Table1_4')
}
\description{
Source: Economic Report of the President, 2007, Table B–110, p. 356
}
\keyword{datasets}
