% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_5.R
\docType{data}
\name{Table1_5}
\alias{Table1_5}
\title{Table 1.5
Seasonally Adjusted M1 Supply: 1959:01 to 1999:07 (billions of dollars)}
\format{
\itemize{
\item \strong{Year:Month}
\item \strong{M1}
}
}
\usage{
data('Table1_5')
}
\description{
Source: Board of Governors, Federal Reserve Bank, USA
}
\keyword{datasets}
