% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_6.R
\docType{data}
\name{Table1_6}
\alias{Table1_6}
\title{Table 1.6
Impact of Advertising Expenditure}
\format{
\itemize{
\item \strong{Firm}
\item \strong{Impressions: } in Millions
\item \strong{Expenditure: } in Millions of 1983 dollars
}
}
\usage{
data('Table1_6')
}
\description{
Source: http://lib.stat.cmu.edu/DASL/Datafiles/tvadsdat.html.
}
\keyword{datasets}
