% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table20_1.R
\docType{data}
\name{Table20_1}
\alias{Table20_1}
\title{Table 20_1}
\format{
\itemize{
\item \strong{Observation}
\item \strong{Q: }Index of Crop Production (1996 = 100)
\item \strong{P: }Index of Crop Prices Received by Farmers (1990 to 1992 = 100)
\item \strong{X: }Real per Capita Personal Consumption Expenditure
}
}
\usage{
data('Table20_1')
}
\description{
Source: Economic Report of the President, 2007. Data on Q (Table B-99), on P (Table B-101), and on X (Table B-31).
}
\keyword{datasets}
