% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table20_10.R
\docType{data}
\name{Table20_10}
\alias{Table20_10}
\title{Table 20_10}
\format{
\itemize{
\item \strong{Educ}
\item \strong{South}
\item \strong{Sex}
\item \strong{Exp}
\item \strong{Union}
\item \strong{Wage}
\item \strong{Age}
\item \strong{Race}
\item \strong{Occup}
\item \strong{Sector}
\item \strong{Marr}
}
}
\usage{
data('Table20_10')
}
\description{
Source: Data can be found on the Web, at http://lib.stat.cmu.edu/datasets/cps_85_wages
}
\keyword{datasets}
