% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table20_2.R
\docType{data}
\name{Table20_2}
\alias{Table20_2}
\title{Table 20_2}
\format{
\itemize{
\item \strong{Observation}
\item \strong{Y1: }GDP = gross domestic product (billions of chained 2000 dollars)
\item \strong{Y2: }M2 = M2 money supply (billions of dollars)
\item \strong{X1: }GPDI = gross private domestic investment (billions of chained 2000 dollars).
\item \strong{X2: }FEDEXP = Federal government expenditure (billions of dollars).
\item \strong{X3: }TB6 = 6-month Treasury bill rate (%).
}
}
\usage{
data('Table20_2')
}
\description{
Smoking and Lung Cancer
Source: http://lib.stat.cmu.edu/ DASL/Datafiles/SmokingandCancer.html
}
\keyword{datasets}
