% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table20_5.R
\docType{data}
\name{Table20_5}
\alias{Table20_5}
\title{Table 20_5}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{C.: } consumption expenditure
\item \strong{P: }profits
\item \strong{W: } private wage bill
\item \strong{I: } investment expenditure
\item \strong{K.1: }capital stock
\item \strong{X: }income after tax (?)
\item \strong{W.: }government wage bill
\item \strong{G: }government expenditure
\item \strong{T: }taxes
}
}
\usage{
data('Table20_5')
}
\description{
Source: These data are taken from G. S. Maddala, Econometrics, McGraw-Hill, New York, 1977, p. 238
}
\keyword{datasets}
