% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table22_5.R
\docType{data}
\name{Table22_5}
\alias{Table22_5}
\title{Table 22_5}
\format{
\itemize{
\item \strong{TIME}
\item \strong{HWI: }Help-Wanted Index
\item \strong{UNRATE: }Unemployment Rate
}
}
\usage{
data('Table22_5')
}
\description{
Source: Marc A. Giammatteo (West Point, Class of 2000), “The Relationship between the Help Wanted Index and the Unemployment Rate,” unpublished term paper. (Notations altered to conform to our notation.)
}
\keyword{datasets}
