% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table22_6.R
\docType{data}
\name{Table22_6}
\alias{Table22_6}
\title{Table 22_6}
\format{
\itemize{
\item \strong{Yen.exchange}
\item \strong{Time}
\item \strong{Year}
\item \strong{Month}
}
}
\usage{
data('Table22_6')
}
\description{
Source: See the page 797 to more information
}
\keyword{datasets}
