% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table22_7.R
\docType{data}
\name{Table22_7}
\alias{Table22_7}
\title{Table 22_7}
\format{
\itemize{
\item \strong{Year}
\item \strong{Month}
\item \strong{TB3M: } three-month Treasury bill rates from January 1, 1982, to March 2008, for a total of 315 monthly observations
\item \strong{TB6M: }six-month Treasury bill rates from January 1, 1982, to March 2008, for a total of 315 monthly observations
}
}
\usage{
data('Table22_7')
}
\description{
Source: See the page 800 to more information
}
\keyword{datasets}
