% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_10.R
\docType{data}
\name{Table2_10}
\alias{Table2_10}
\title{Table 2.10}
\format{
\itemize{
\item \strong{Family.Income}
\item \strong{Middle.Income} Middle Income in the interval
\item \strong{Number.of.test.takers}
\item \strong{CR_Mean: } Critical Reading Mean
\item \strong{CR_SD: } Critical Reading Standart Deviation
\item \strong{M_Mean: } Mathematics Mean
\item \strong{M_SD: } Mathematics Standart Deviation
\item \strong{W_Mean: } Writing Mean
\item \strong{W_SD: } Writing Standart Deviation
}
}
\usage{
data('Table2_10')
}
\description{
SAT Reasoning Test Classified by Family Income
}
\details{
Source: College Board, 2007 College-Bound Seniors, Table 11.
}
\keyword{datasets}
