% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_4.R
\docType{data}
\name{Table2_4}
\alias{Table2_4}
\title{Table 2.4
A Random Sample from the Population of Table 2.1}
\format{
\itemize{
\item \strong{Y}
\item \strong{X}
}
}
\usage{
data('Table2_4')
}
\description{
Table 2.4
A Random Sample from the Population of Table 2.1
}
\keyword{datasets}
