% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_5.R
\docType{data}
\name{Table2_5}
\alias{Table2_5}
\title{Table 2.5
Another Random Sample from the Population of Table 2.1}
\format{
\itemize{
\item \strong{Y}
\item \strong{X}
}
}
\usage{
data('Table2_5')
}
\description{
Table 2.5
Another Random Sample from the Population of Table 2.1
}
\keyword{datasets}
