% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_6.R
\docType{data}
\name{Table2_6}
\alias{Table2_6}
\title{Table 2.6
Mean Hourly Wage by Education}
\format{
\itemize{
\item \strong{Years of Schooling}
\item \strong{Mean Wage}
\item \strong{Number of People}
}
}
\usage{
data('Table2_6')
}
\description{
Source: Arthur S. Goldberger, Introductory Econometrics, Harvard University Press, Cambridge, Mass., 1998, Table 1.1, p. 5 (adapted).
}
\keyword{datasets}
