% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_7.R
\docType{data}
\name{Table2_7}
\alias{Table2_7}
\title{Table 2.7
Labor Force Participation Data for U.S. for 1980 to 2006}
\format{
\itemize{
\item \strong{Years}
\item \strong{TPFTCH : }Civilian labor force participation rate, male (%), Table B-39, p. 277
\item \strong{TPFTCM : }Civilian labor force participation rate, female (%), Table B-39, p. 277
\item \strong{TDCH : } Civilian unemployment rate, male (%) Table B-42, p. 280
\item \strong{TDCM : } Civilian unemployment rate, female (%) Table B-42, p. 280.
\item \strong{GMH82  : }Average hourly earnings (1982 dollars), Table B-47, p. 286.
\item \strong{GMH: }Average hourly earnings (current dollars), Table B-47, p. 286
}
}
\usage{
data('Table2_7')
}
\description{
Source: Economic Report of the President, 2007.
}
\keyword{datasets}
