% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_8.R
\docType{data}
\name{Table2_8}
\alias{Table2_8}
\title{Table 2.8
Food and Total Expenditure (Rupees)}
\format{
\itemize{
\item \strong{Observation}
\item \strong{FOODEXP  : }FODD EXPENDITURE
\item \strong{TOTALEXP : }TOTAL EXPENDITURE
}
}
\usage{
data('Table2_8')
}
\description{
Source: Chandan Mukherjee, Howard White, and Marc Wuyts, Econometrics and Data Analysis for Developing Countries, Routledge, New York, 1998, p. 457
}
\keyword{datasets}
