% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2_9.R
\docType{data}
\name{Table2_9}
\alias{Table2_9}
\title{Table 2.9
Total Group Mean SAT Reasoning Test Scores: CollegeBound Seniors, 1972–2007}
\format{
\itemize{
\item \strong{Year}
\item \strong{CR_MALE_R : } MALE CR IN CRITICAL READING
\item \strong{CR_FEMALE_R : } FEMALE CR IN CRITICAL READING
\item \strong{CR_MALE_M : } MALE CR IN MATHEMATICS
\item \strong{CR_FEMALE_M : } FEMALE CR IN MATHEMATICS
}
}
\usage{
data('Table2_9')
}
\description{
Source: College Board, 2007.
}
\keyword{datasets}
