% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_2.R
\docType{data}
\name{Table3_2}
\alias{Table3_2}
\title{Table 3.2
Raw Data Based on Table 2.6}
\format{
\itemize{
\item \strong{Mean.Hourly.Wage}
\item \strong{Years.of.Schooling}
}
}
\usage{
data('Table3_2')
}
\description{
Table 3.2
Raw Data Based on Table 2.6
}
\keyword{datasets}
