% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_3.R
\docType{data}
\name{Table3_3}
\alias{Table3_3}
\title{Table 3.3
Number of Cellular Phone Subscribers per Hundred Persons and Number of Personal Computers per 100 Persons and Per Capita Income in Selected Countries for 2003}
\format{
\itemize{
\item \strong{Country}
\item \strong{Cellphone}
\item \strong{PCs}
\item \strong{Per Capita Income ($)}
\item \strong{Ln.Cellphone} Natural Log of the variable Cellphone
\item \strong{Ln.PCs} Natural Log of the variable PCs
\item \strong{Ln.Income ($)} Natural Log of the variable Per Capita Income
}
}
\usage{
data('Table3_3')
}
\description{
Source: Statistical Abstract of the United States, 2006, Table 1364 for data on cell phones and computers and Table 1327 for purchasingpower adjusted per capita income
}
\note{
Note: The data on cell phones and personal computers are per 100 persons.
}
\keyword{datasets}
