% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_5.R
\docType{data}
\name{Table3_5}
\alias{Table3_5}
\title{Table 3.5}
\format{
\itemize{
\item \strong{Student}
\item \strong{Partial}
\item \strong{Final}
}
}
\usage{
data('Table3_5')
}
\description{
Table 3.5
}
\keyword{datasets}
