% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_6.R
\docType{data}
\name{Table3_6}
\alias{Table3_6}
\title{Table 3.6
Productivity and Related Data, Business Sector 1960 to 2005 (Index numbers, 1992 =100; quarterly data seasonally adjusted)}
\format{
\itemize{
\item \strong{ANO: } year
\item \strong{OUT_BUS} Output per hour for all persons on Business Sector
\item \strong{OUT_NBUS} Output per hour for all persons on Nonfarm Business Sector
\item \strong{COM_BUS} Real compensation per hour for all persons on Business Sector
\item \strong{COM_NBUS} Real compensation per hour for all persons on Nonfarm Business Sector
}
}
\usage{
data('Table3_6')
}
\description{
Source: Economic Report of the President, 2007, Table 49
}
\note{
Output refers to real gross domestic product in the sector. Wages and salaries of employees plus employers’ contributions for social insurance and private benefit plans.Hourly compensation divided by the consumer price index for all urban consumers for recent quarters.
}
\keyword{datasets}
