% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5_10.R
\docType{data}
\name{Table5_10}
\alias{Table5_10}
\title{Table 5.10}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{CPI: } Consumer Price Index
\item \strong{PPI: } Wholesale Price Index, also called Produce Price Index
}
}
\usage{
data('Table5_10')
}
\description{
CPI and PPI, USA,1980–2006
Source: Economic Report of the President, 2007, Tables B-62 and B-65.
}
\keyword{datasets}
