% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5_11.R
\docType{data}
\name{Table5_11}
\alias{Table5_11}
\title{Table 5.11}
\format{
\itemize{
\item \strong{Occupational_Group}
\item \strong{Smoking}
\item \strong{Mortality}
}
}
\usage{
data('Table5_11')
}
\description{
Smoking and Lung Cancer
Source: http://lib.stat.cmu.edu/ DASL/Datafiles/SmokingandCancer.html
}
\keyword{datasets}
