% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5_5.R
\docType{data}
\name{Table5_5}
\alias{Table5_5}
\title{Table 5.5}
\format{
\itemize{
\item \strong{Year}
\item \strong{SALARY :} Nominal Gross Domestic Product, 1959 to 2005
\item \strong{SPENDING :} Real Gross Domestic Product, prices seted in 2000
}
}
\usage{
data('Table5_5')
}
\description{
Average Salary and Per Pupil Spending (dollars), 1985
}
\details{
Source: National Education Association, as reported by Albuquerque Tribune, Nov. 7, 1986#'
}
\keyword{datasets}
