% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5_6.R
\docType{data}
\name{Table5_6}
\alias{Table5_6}
\title{Table 5.6}
\format{
\itemize{
\item \strong{Year}
\item \strong{GNP: } Gross Nacional Product
\item \strong{M1: } Currency + Demand deposits + Travelers checks and other checkable deposits (OCDs)
\item \strong{M2: } M1 + Overnight RPs and Eurodollars + MMMF (Money market mutual fund) balances + MMDAs (Money market deposit accounts) + Savings and small deposits
\item \strong{M3: } M2 + Large time deposits + Term RPs + Institutional MMMF
\item \strong{L: } M3 + Other liquid assets
}
}
\usage{
data('Table5_6')
}
\description{
GNP and Four Measures of Money Stock
Source: Economic Report of the President, 1985, GNP data from Table B-1, p. 232; money stock data from Table B-61, p. 303.
}
\keyword{datasets}
