% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table5_9.R
\docType{data}
\name{Table5_9}
\alias{Table5_9}
\title{Table 5.9}
\format{
\itemize{
\item \strong{COUNTRY}
\item \strong{BMACLC :} Big Mac price in local currency
\item \strong{BMAC. :} Big Mac price in dollars
\item \strong{IPPP: } Implied PPP of the dollar (PPP is Purchasing power parity)
\item \strong{ADER: } Actual dollar exchange Rate in Jan 31
\item \strong{U.OVAD: } Under/Over valuation against the dollar (%)
}
}
\usage{
data('Table5_9')
}
\description{
The Hamburger Standard
Source: McDonald’s; The Economist, February 1, 2007
}
\keyword{datasets}
