% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_1.R
\docType{data}
\name{Table6_1}
\alias{Table6_1}
\title{Table 6.1}
\format{
\itemize{
\item \strong{OBS} Year and Month
\item \strong{Y: } Excess returns (%) on an index of 104 stocks
\item \strong{X: } Excess returns (%) on the overall stocks
}
}
\usage{
data('Table6_1')
}
\description{
Data on excess returns Yt (%) on an index of 104 stocks in the sector of cyclical consumer goods and excess returns Xt (%) on the overall stock market index for the U.K. for the monthly data for the period 1980–1999, for a total of 240 observations.5 Excess return refers to return in excess of return on a riskless asset (see the CAPM model)
}
\details{
Source: These data, originally obtained from DataStream databank, are reproduced from Christiaan Heij et al., Econometrics Methods with Applications in Business and Economics, Oxford University Press, Oxford, U.K., 2004.
}
\keyword{datasets}
