% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_10.R
\docType{data}
\name{Table6_10}
\alias{Table6_10}
\title{Table 6.10}
\format{
\itemize{
\item \strong{Exp: }Advertising expenditure (£, millions)
\item \strong{Adexp: }Total consumer expenditure (£, millions)
\item \strong{Ratio: } Ratio
\item \strong{ln.Exp: }ln.Exp
\item \strong{ln.Adexp: }ln.Adexp
\item \strong{X1/Exp: } 1/Exp
\item \strong{X1/Adexp: } 1/Adexp
}
}
\usage{
data('Table6_10')
}
\description{
Advertising Expenditure and Total Expenditure (in £ millions) for 29 Product Categories in the U.K.
Source: http://www.Economicswebinstitute.org/ecdata.htm
}
\keyword{datasets}
