% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_2.R
\docType{data}
\name{Table6_2}
\alias{Table6_2}
\title{Table 6.2}
\format{
\itemize{
\item \strong{Year}
\item \strong{GPDIBL: }gross private domestic investment, billions of 2000 dollars
\item \strong{GPDIM: }gross private domestic investments, millions of 2000 dollars
\item \strong{GDPB: }gross domestic product, billions of 2000 dollars.
\item \strong{GDPM: }gross domestic product, millions of 2000 dollars
}
}
\usage{
data('Table6_2')
}
\description{
Gross Private Domestic Investment and GDP, United States, 1990 to 2005 (Billions of chained  [2000] dollars, except as noted; quarterly data at seasonally adjusted annual rates)
Source: Economic Report of the President, 2007, Table B2, p. 328
}
\keyword{datasets}
