% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_3.R
\docType{data}
\name{Table6_3}
\alias{Table6_3}
\title{Table 6.3}
\format{
\itemize{
\item \strong{Year.or.quarter}
\item \strong{EXPSERVICES: }expenditure on services, billions of 2000 dollars
\item \strong{EXPDUR: }expenditure on durable goods, billions of 2000 dollars
\item \strong{EXPNONDUR: }expenditure on nondurable goods, billions of 2000 dollars
\item \strong{PCEXP: }total personal consumption expenditure, billions of 2000 dollars
}
}
\usage{
data('Table6_3')
}
\description{
Total Personal Expenditure and Categories (Billions of chained [2000] dollars; quarterly data at seasonally adjusted annual rates)
Sources: Department of Commerce, Bureau of Economic Analysis. Economic Report of the President, 2007, Table B-17, p. 347
}
\keyword{datasets}
