% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_4.R
\docType{data}
\name{Table6_4}
\alias{Table6_4}
\title{Table 6.4}
\format{
\itemize{
\item \strong{Occupational_Group}
\item \strong{Smoking}
\item \strong{Mortality}
}
}
\usage{
data('Table6_4')
}
\description{
Fertility and Other Data for 64 Countries
Source: Chandan Mukherjee, Howard White, and Marc Whyte, Econometrics and Data Analysis for Developing Countries, Routledge, London, 1998, p. 456
}
\keyword{datasets}
