% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_5.R
\docType{data}
\name{Table6_5}
\alias{Table6_5}
\title{Table 6.5}
\format{
\itemize{
\item \strong{Year}
\item \strong{INFLRATE: } Inflation Rate
\item \strong{UNRATE: } Unemployment Rate
}
}
\usage{
data('Table6_5')
}
\description{
Inflation Rate and Unemployment Rate, United States, 1960–2006 (For all urban consumers; 1982–1984 = 100, except as noted)
Source: Economic Report of the President, 2007, Table B-60, p. 399, for CPI changes and Table B-42, p. 376, for the unemployment rate.
}
\keyword{datasets}
