% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_7.R
\docType{data}
\name{Table6_7}
\alias{Table6_7}
\title{Table 6.7}
\format{
\itemize{
\item \strong{Y}
\item \strong{X}
}
}
\usage{
data('Table6_7')
}
\description{
Source: Adapted from J. Johnston, Econometric Methods, 3d ed., McGraw-Hill, New York, 1984, p. 87. Actually this is taken from an econometric examination of Oxford University in 1975.
}
\keyword{datasets}
