% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_8.R
\docType{data}
\name{Table6_8}
\alias{Table6_8}
\title{Table 6.8}
\format{
\itemize{
\item \strong{COUNTRY}
\item \strong{SAVRATE: }Savings as a ratio of GDP
\item \strong{INVRATE: }Investment expenditure as a ratio of GDP
}
}
\usage{
data('Table6_8')
}
\description{
Relationship between investment rate and savings rate
Source: Martin Feldstein and Charles Horioka, “Domestic Saving and International Capital Flows,” Economic Journal, vol. 90, June 1980, pp. 314–329. Data reproduced from Michael P. Murray, Econometrics: A Modern Introduction, Addison-Wesley, Boston, 2006.
}
\keyword{datasets}
