% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table6_9.R
\docType{data}
\name{Table6_9}
\alias{Table6_9}
\title{Table 6.9}
\format{
\itemize{
\item \strong{wfood: } budget share for food expenditure
\item \strong{wfuel: }budget share for fuel expenditure
\item \strong{wcloth: }budget share for clothing expenditure
\item \strong{walc: } budget share for alcohol expenditure
\item \strong{wtrans: }budget share for transportation expenditure
\item \strong{wother: }budget share for other expenditures
\item \strong{totexp: }total household expenditure
\item \strong{income: }total net household income
\item \strong{age: }age of household head
\item \strong{nk: }number of children
\item \strong{food: } value of food expenditure
\item \strong{fuel: }value of fuel expenditure
\item \strong{cloth: }value of clothing expenditure
\item \strong{alc: }value of alcohol expenditure
\item \strong{trans: }value of transportation expenditure
\item \strong{other: }value of other expenditures

}
}
\usage{
data('Table6_9')
}
\description{
Sample of 1,519 households drawn from the 1980–1982 British Family Expenditure Surveys
Source: The data are from Richard Blundell and Krishna Pendakur, “Semiparametric Estimation and Consumer Demand,” Journal of Applied Econometrics, vol. 13, no. 5, 1998, pp. 435–462. Data reproduced from R. Carter Hill, William E. Griffiths, and George G. Judge, Undergraduate Econometrics, 2d ed., John Wiley & Sons, New York, 2001.
}
\keyword{datasets}
