% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_1.R
\docType{data}
\name{Table7_1}
\alias{Table7_1}
\title{Table 7_1}
\format{
\itemize{
\item \strong{Year}
\item \strong{Y: }Cups per Person, per Day
\item \strong{X: }$ per lb
}
}
\usage{
data('Table7_1')
}
\description{
U.S. Coffee Consumption (Y ) in Relation to Average Real Retail Price (X),* 1970–1980
Source: The data for Y are from Summary of National Coffee Drinking Study, Data Group, Elkins Park, Penn., 1981; and the data on nominal X (i.e., X in current prices) are from Nielsen Food Index, A. C. Nielsen, New York, 1981.
I am indebted to Scott E. Sandberg for collecting the data.
}
\note{
The nominal price was divided by the Consumer Price Index (CPI) for food and beverages, 1967 = 100.
}
\keyword{datasets}
