% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_11.R
\docType{data}
\name{Table7_11}
\alias{Table7_11}
\title{Table 7_11}
\format{
\itemize{
\item \strong{obs} Year
\item \strong{OUTPUT: }Billions of Drachmas at constant 1970 prices
\item \strong{CAPITAL: }
\item \strong{LABOR: }Thousands of workers per year.
\item \strong{CLRATIO: } Capital to labor ratio
}
}
\usage{
data('Table7_11')
}
\description{
Greek Industrial Sector
Source: I am indebted to George K. Zestos of Christopher Newport University, Virginia, for these data.
}
\keyword{datasets}
