% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_12.R
\docType{data}
\name{Table7_12}
\alias{Table7_12}
\title{Table 7_12}
\format{
\itemize{
\item \strong{Year}
\item \strong{C: }real consumption expenditures in billions of chained 1996 dollars.
\item \strong{Yd: } real personal disposable income in billions of chained 1996 dollars
\item \strong{wealth: } real wealth in billions of chained 1996 dollars
\item \strong{interest: }nominal annual yield on 3-month Treasury securities–inflation rate (measured by the annual % change in annual chained price index).
}
}
\usage{
data('Table7_12')
}
\description{
Real Consumption Expenditure, Real Income, Real Wealth, and Real Interest Rates for the U.S., 1947 to 2000
Sources: C, Yd, and quarterly and annual chain-type price indexes (1996 = 100): Bureau of Economic Analysis, U.S. Department of Commerce (http://www.bea.doc.gov/bea/dn1.htm). Nominal annual yield on 3-month Treasury securities:  Economic Report of the President, 2002. Nominal wealth = end-ofyear nominal net worth of households and nonprofits (from Federal Reserve flow of funds data: http://www. federalreserve.gov).
}
\keyword{datasets}
