% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_13.R
\docType{data}
\name{Table7_13}
\alias{Table7_13}
\title{Table 7_13}
\format{
\itemize{
\item \strong{Date}
\item \strong{time}
\item \strong{Close}
}
}
\usage{
data('Table7_13')
}
\description{
Estimating Qualcomm stock prices
Source: The full data can be found on the textbook’s website in Table 7.13
}
\keyword{datasets}
