% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_3.R
\docType{data}
\name{Table7_3}
\alias{Table7_3}
\title{Table 7_3}
\format{
\itemize{
\item \strong{COUNTRY:} Area
\item \strong{produto.valor.agregado.em.milhares..: }Output Value Added (thousands of $)
\item \strong{insumo.trabalho.horas.trabalhadas.em.milhares: }Labor Input Worker Hrs (thousands)
\item \strong{insumo.capital.despesa.capital.em.milhares..: }Capital Input Capital Expenditure (thousands)
}
}
\usage{
data('Table7_3')
}
\description{
Value Added, Labor Hours, and Capital Input in the Manufacturing Sector of the U.S., 2005
Source: 2005 Annual Survey of Manufacturers, Sector 31: Supplemental Statistics for U.S.
}
\keyword{datasets}
