% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_4.R
\docType{data}
\name{Table7_4}
\alias{Table7_4}
\title{Table 7_4}
\format{
\itemize{
\item \strong{X: }Output
\item \strong{y: }Total Cost
}
}
\usage{
data('Table7_4')
}
\description{
Total Cost (Y) and Output (X)
}
\keyword{datasets}
