% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_5.R
\docType{data}
\name{Table7_5}
\alias{Table7_5}
\title{Table 7_5}
\format{
\itemize{
\item \strong{Y}
\item \strong{X2}
\item \strong{X3}
}
}
\usage{
data('Table7_5')
}
\description{
Table 7_5
}
\keyword{datasets}
