% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_6.R
\docType{data}
\name{Table7_6}
\alias{Table7_6}
\title{Table 7_6}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{Y: }quantity of roses sold, dozens
\item \strong{X2: }average wholesale price of roses, $/dozen
\item \strong{X3: }average wholesale price of carnations, $/dozen
\item \strong{X4: }average weekly family disposable income, $/week
\item \strong{X5: } the trend variable taking values of 1, 2, and so on, for the period 1971–III to 1975–II in the Detroit metropolitan area
}
}
\usage{
data('Table7_6')
}
\description{
Quarterly Demand for Roses in Metro Detroit Area, from 1971-III to 1975-II
Source: I am indebted to Joe Walsh for collecting these data from a major wholesaler in the Detroit metropolitan area and subsequently processing them
}
\keyword{datasets}
