% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_7.R
\docType{data}
\name{Table7_7}
\alias{Table7_7}
\title{Table 7_7}
\format{
\itemize{
\item \strong{Y: }Thousands of Wildcats
\item \strong{X2: } Per Barrel Price, Constant $
\item \strong{X3: } Domestic Output (millions of barrels per day)
\item \strong{X4: } GNP, Constant $ Billions
\item \strong{X5: } Time 1948 = 1
}
}
\usage{
data('Table7_7')
}
\description{
Wildcat Activity
Source: Energy Information Administration, 1978 Report to Congress
}
\keyword{datasets}
