% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_8.R
\docType{data}
\name{Table7_8}
\alias{Table7_8}
\title{Table 7_8}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{Y: } Defense Budget Outlays
\item \strong{X2: } GNP
\item \strong{X3: } US Military Sales/Assistance
\item \strong{X4: } Aerospace Industry Sales
\item \strong{X5: } Conflicts 100,000+
}
}
\usage{
data('Table7_8')
}
\description{
U.S. Defense Budget Outlays, 1962–1981
Source: These data were collected by Albert Lucchino from various government publications
}
\keyword{datasets}
