% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table7_9.R
\docType{data}
\name{Table7_9}
\alias{Table7_9}
\title{Table 7_9}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{Y: }per capita consumption of chickens, lb
\item \strong{X2: }real disposable income per capita, $
\item \strong{X3: }real retail price of chicken per lb, ¢
\item \strong{X4: }real retail price of pork per lb, ¢
\item \strong{X5: }real retail price of beef per lb, ¢
\item \strong{X6: }composite real price of chicken substitutes per lb, ¢, which is a weighted average of the real retail prices per lb of pork and beef, the weights being the relative consumptions of beef and pork in total beef and pork consumption
}
}
\usage{
data('Table7_9')
}
\description{
Demand for Chicken in the U.S., 1960–1982
Source: Data on Y are from Citibase and on X2 through X6 are from the U.S. Department of Agriculture. I am indebted to Robert J. Fisher for collecting the data and for the statistical analysis
}
\note{
The real prices were obtained by dividing the nominal prices by the Consumer Price Index for food
}
\keyword{datasets}
