% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8_10.R
\docType{data}
\name{Table8_10}
\alias{Table8_10}
\title{Table 8_10}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{X2: }GNP
\item \strong{X3: }Housing Starts
\item \strong{X4: }Unemployment %
\item \strong{X5: }Prime Rate Lag, 6 mos.
\item \strong{X6: }Customer Line, Gains %
\item \strong{Y: }Annual Sales (MPF)
}
}
\usage{
data('Table8_10')
}
\description{
The demand for cable
Source: I am indebted to Daniel J. Reardon for collecting and processing the data
}
\keyword{datasets}
