% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8_11.R
\docType{data}
\name{Table8_11}
\alias{Table8_11}
\title{Table 8_11}
\format{
\itemize{
\item \strong{Year}
\item \strong{Saving}
\item \strong{Income}
}
}
\usage{
data('Table8_11')
}
\description{
Savings and Personal Disposable Income (billions of dollars), United States,1970–1995
Source: Department of Commerce, Bureau of Economic Analysis
}
\keyword{datasets}
