% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8_8.R
\docType{data}
\name{Table8_8}
\alias{Table8_8}
\title{Table 8_8}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{GDP: }Millions of 1960 pesos.
\item \strong{Employment: }Thousands of people
\item \strong{CAPITAL: }Millions of 1960 pesos.
}
}
\usage{
data('Table8_8')
}
\description{
Real GDP, Employment, and Real Fixed Capital—Mexico
Source: Victor J. Elias, Sources of Growth: A Study of Seven Latin American Economies, International Center for Economic Growth, ICS Press, San Francisco, 1992. Data from Tables E5, E12, and E14.
}
\keyword{datasets}
