% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table8_9.R
\docType{data}
\name{Table8_9}
\alias{Table8_9}
\title{Table 8_9}
\format{
\itemize{
\item \strong{YEAR}
\item \strong{SAVINGS}
\item \strong{INCOME}
}
}
\usage{
data('Table8_9')
}
\description{
Savings and Personal Disposable Income (billions of dollars), United States,1970–1995
Source: Economic Report of the President, 1997, Table B-28, p. 332
}
\keyword{datasets}
