% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_1.R
\docType{data}
\name{Table9_1}
\alias{Table9_1}
\title{Table 9_1}
\format{
\itemize{
\item \strong{state }
\item \strong{Salary }
\item \strong{Spending }
\item \strong{D2: }1 for states in the Northeast and North Central; 0 otherwise.
\item \strong{D3: }1 for states in the South; 0 otherwise.
}
}
\usage{
data('Table9_1')
}
\description{
Average Salary of Public School Teachers by State, 2005–2006
Source: National Educational Association, as reported in 2007
}
\keyword{datasets}
