% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_3.R
\docType{data}
\name{Table9_3}
\alias{Table9_3}
\title{Table 9_3}
\format{
\itemize{
\item \strong{DISH: }dishwashers
\item \strong{DISP: }garbage disposers
\item \strong{FRIG: }refrigerators
\item \strong{WASH: }washing machines
\item \strong{DUR: }durable goods expenditure (billions of 1982 dollars)
}
}
\usage{
data('Table9_3')
}
\description{
Quarterly Data on Appliance Sales (in thousands) and Expenditure on Durable Goods (1978–I to 1985–IV)
Source: Business Statistics and Survey of Current Business, Department of Commerce (various issues).
}
\keyword{datasets}
