% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_4.R
\docType{data}
\name{Table9_4}
\alias{Table9_4}
\title{Table 9_4}
\format{
\itemize{
\item \strong{FRIG: }refrigerator sales, thousands.
\item \strong{DUR: }durable goods expenditure, billions of 1982 dollars.
\item \strong{D2: }1 in the second quarter, 0 otherwise
\item \strong{D3: }1 in the third quarter, 0 otherwise
\item \strong{D4: }1 in the fourth quarter, 0 otherwise
}
}
\usage{
data('Table9_4')
}
\description{
U.S. Refrigerator Sales (thousands), 1978–1985 (quarterly)
Source: Business Statistics and Survey of Current Business, Department of Commerce (various issues)
}
\keyword{datasets}
