% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_6.R
\docType{data}
\name{Table9_6}
\alias{Table9_6}
\title{Table 9_6}
\format{
\itemize{
\item \strong{Total.cost....} Total Cost, Dollars
\item \strong{Output..units.} Output, Units
}
}
\usage{
data('Table9_6')
}
\description{
Hypothetical Data on Output and Total Cost
}
\keyword{datasets}
