% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_7.R
\docType{data}
\name{Table9_7}
\alias{Table9_7}
\title{Table 9_7}
\format{
\itemize{
\item \strong{WI: }weekly wage income in rupees
\item \strong{AGE: }age in years
\item \strong{DE2: }a dummy variable taking a value of 1 for workers with an education level up to primary
\item \strong{DE3: }a dummy variable taking a value of 1 for workers up to a secondary level of education
\item \strong{DE4: }a dummy variable taking a value of 1 for workers with higher than secondary education
\item \strong{DPT: }a dummy variable taking a value of 1 for workers with permanent jobs and a value of 0 for temporary workers
\item \strong{DSEX: }1 for male workers and 0 for female workers
}
}
\usage{
data('Table9_7')
}
\description{
Indian Wage Earners, 1990
Source: http://lib.stat.cmu.edu/ DASL/Datafiles/SmokingandCancer.html
}
\keyword{datasets}
