% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table9_8.R
\docType{data}
\name{Table9_8}
\alias{Table9_8}
\title{Table 9_8}
\format{
\itemize{
\item \strong{YEAR_QUART: } Year and Quarter
\item \strong{UN: } Unemployment Rate %
\item \strong{V: } Job Vacancy Rate %
\item \strong{DUM: } 1 for period beginning in 1966-IV and 0 for before that
\item \strong{DUMV: } V when DUM = 1
}
}
\usage{
data('Table9_8')
}
\description{
Data Matrix for Regression, in Exercise 9.3
Source: Damodar Gujarati, “The Behaviour of Unemployment and Unfilled Vacancies: Great Britain, 1958–1971,” The Economic Journal, vol. 82, March 1972, p. 202
}
\keyword{datasets}
