% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_12.R
\docType{data}
\name{table_10_12}
\alias{table_10_12}
\title{Table 10.12 An Economic Model of the United States}
\format{
A \link[tibble:tibble-package]{tibble} with 14 rows and
5 variables:
\itemize{
\item \strong{year}
\item \strong{y:} consumption
\item \strong{x2:} wage income
\item \strong{x3:} nonwage, nonfarm income
\item \strong{x4:} farm income
}
}
\source{
L. R. Klein and A. S. Goldberger, An Economic Model of the United States,
1929–1952, North Holland Publishing Company, Amsterdam, 1964, p. 131.
}
\usage{
table_10_12
}
\description{
Table 10.12 An Economic Model of the United States
}
\note{
The data for the war years 1942–1944 are missing. The data for other years are billions of 1939 dollars.
}
\keyword{datasets}
