% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_14.R
\docType{data}
\name{table_10_14}
\alias{table_10_14}
\title{Table 10.14 Passenger Car Data}
\format{
A \link[tibble:tibble-package]{tibble} with 16 rows and
7 variables:
\describe{
\item{year}{}
\item{y:}{new passenger cars sold (thousands), seasonally unadjusted}
\item{x2:}{new cars, Consumer Price Index, 1967 = 100, seasonally unadjusted}
\item{x3:}{Consumer Price Index, all items, all urban consumers, 1967 = 100, seasonally unadjusted}
\item{x4:}{the personal disposable income (PDI), billions of dollars, unadjusted for seasonal variation}
\item{x5:}{the interest rate, percent, ﬁnance company paper placed directly}
\item{x6:}{the employed civilian labor force (thousands), unadjusted for seasonal variation.}
}
}
\source{
Business Statistics, 1986, A Supplement to the Current Survey of Business, U.S. Department of Commerce.
}
\usage{
table_10_14
}
\description{
Passenger Car Data
}
\keyword{datasets}
