% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_17.R
\docType{data}
\name{table_10_17}
\alias{table_10_17}
\title{Table 10.17 Updated Longley Data, 1959–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 47 rows and
8 variables:
\itemize{
\item \strong{obs:}
\item \strong{y:} number of people employed, in thousands
\item \strong{x1:} GNP implicit price deﬂator
\item \strong{x2:} GNP, millions of dollars
\item \strong{x3:} number of people unemployed in thousands
\item \strong{x4:} number of people in the armed forces in thousands
\item \strong{x5:} noninstitutionalized population over 16 years of age
\item \strong{x6:} year, equal to 1 in 1959, 2 in 1960, and 47 in 2005
}
}
\source{
Department of Labor, Bureau of Labor Statistics and \url{http://siadapp.dmdc.osd.mil/personnel/MILITARY/Miltop.htm}
}
\usage{
table_10_17
}
\description{
Updated Longley Data, 1959–2005
}
\keyword{datasets}
