% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_18.R
\docType{data}
\name{table_10_18}
\alias{table_10_18}
\title{Table 10.18 Chemicals in Cheeses}
\format{
A \link[tibble:tibble-package]{tibble} with 30 rows and
5 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{taste}
\item \strong{acetic:} the natural logarithm of concentration of acetic acid
\item \strong{h2s:} the natural logarithm of concentration of hydrogen sulfide
\item \strong{lactic:}
}
}
\source{
\url{http://lib.stat.cmu.edu/DASL/Dataﬁles/Cheese.html}
}
\usage{
table_10_18
}
\description{
As cheese ages, several chemical processes take place that determine the taste
of the ﬁnal product. The data pertain to concentrations of various chemicals
in a sample of 30 mature cheddar cheeses and subjective measures of taste for each sample.
}
\keyword{datasets}
