% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_3.R
\docType{data}
\name{table_10_3}
\alias{table_10_3}
\title{Table 10.3 Hypothetical Data on Y, X2, and X3}
\format{
A \link[tibble:tibble-package]{tibble} with 5 rows and
3 variables:
\itemize{
\item \strong{y}
\item \strong{x2}
\item \strong{x3}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 332.
}
\usage{
table_10_3
}
\description{
Table 10.3 Hypothetical Data on Y, X2, and X3
}
\keyword{datasets}
