% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_5.R
\docType{data}
\name{table_10_5}
\alias{table_10_5}
\title{Table 10.5 Hypothetical Data on Consumption Expenditure Y, Income X2, and Wealth X3}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
3 variables:
\itemize{
\item \strong{y:} consumption expenditure
\item \strong{x2:} income
\item \strong{x3:} wealth
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 333.
}
\usage{
table_10_5
}
\description{
Hypothetical Data on Consumption Expenditure Y, Income X2, and Wealth X3
}
\keyword{datasets}
