% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_7.R
\docType{data}
\name{table_10_7}
\alias{table_10_7}
\title{Table 10.7 U.S. Consumption Expenditure for the Period 1947–2000}
\format{
A \link[tibble:tibble-package]{tibble} with 54 rows and
5 variables:
\describe{
\item{year}{}
\item{c:}{consume expenditure}
\item{yd:}{real disposable personal income}
\item{w:}{real wealth}
\item{i:}{real interest rate}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 335. See Table 7.12.
}
\usage{
table_10_7
}
\description{
Data on real consumption expenditure (C), real disposable personal income (Yd),
real wealth (W), and real interest rate (I) for the United States for the period 1947–2000
}
\keyword{datasets}
