% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_8.R
\docType{data}
\name{table_10_8}
\alias{table_10_8}
\title{Table 5.11 Longley Data}
\format{
A \link[tibble:tibble-package]{tibble} with 16 rows and
8 variables:
\itemize{
\item \strong{year}
\item \strong{y:} number of people employed, in thousands
\item \strong{x1:} GNP implicit price deflator
\item \strong{x2:} GNP, millions of dollar2
\item \strong{x3:} number of people unemployed, in thousands
\item \strong{x4:} number of people in the armed forces
\item \strong{x5:} noninstitutionalized population over 14 years of age
\item \strong{time:} year, equal to 1 in 1947, 2 in 1948, and 16 in 1962
}
}
\source{
J. Longley, “An Appraisal of Least-Squares Programs from the Point of
the User,” Journal of the American Statistical Association, vol. 62, 1967, pp. 819–841.
}
\usage{
table_10_8
}
\description{
The data are time series for the years 1947–1962
}
\keyword{datasets}
