% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_3.R
\docType{data}
\name{table_11_3}
\alias{table_11_3}
\title{Table 11.3 Hypothetical Data on Consumption Expenditure Y($) and Income X($) to Illustrate the Goldfeld–Quandt Test}
\format{
A \link[tibble:tibble-package]{tibble} with 30 rows and
4 variables:
\itemize{
\item \strong{y:} consumption expenditure, $
\item \strong{x:} income, $
\item \strong{ry:} consumption expenditure ranked by x values
\item \strong{rx:} ranked income
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 384.
}
\usage{
table_11_3
}
\description{
Table 11.3 Hypothetical Data on Consumption Expenditure Y($) and Income X($) to Illustrate the Goldfeld–Quandt Test
}
\keyword{datasets}
