% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_5.R
\docType{data}
\name{table_11_5}
\alias{table_11_5}
\title{Table 11.5 Sales and Employment for Companies Performing Industrial R&D in the United States, by Industry, 2005}
\format{
A \link[tibble:tibble-package]{tibble} with 14 rows and
5 variables:
\itemize{
\item \strong{obs:} observations
\item \strong{industry}
\item \strong{sales:} millions of dollars
\item \strong{rd:} millions of dollars
\item \strong{profits:} millions of dollars
}
}
\source{
National Science Foundation, Division of Science Resources Statistics,
Survey of Industrial Research and Development: 2005 and the U.S. Census Bureau
Annual Survey of Manufacturers, 2005.
}
\usage{
table_11_5
}
\description{
Sales and Employment for Companies Performing Industrial R&D in the United States, by Industry, 2005
}
\keyword{datasets}
