% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_7.R
\docType{data}
\name{table_11_7}
\alias{table_11_7}
\title{Table 11.7 Passenger Car Mileage Data}
\format{
A \link[tibble:tibble-package]{tibble} with 81 rows and
6 variables:
\describe{
\item{obs:}{car observation number (names of cars not disclosed)}
\item{mpg:}{average miles per gallon}
\item{sp:}{top speed, miler per hour}
\item{hp:}{engine horsepower}
\item{vol:}{cubic feet of cab space}
\item{wt:}{vehicle weight, hundreds of pounds}
}
}
\source{
U.S. Environmental Protection Agency, 1991, Report EPA /AA /CTAB/91-02.
}
\usage{
table_11_7
}
\description{
Passenger Car Mileage Data
}
\keyword{datasets}
