% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_8.R
\docType{data}
\name{table_11_8}
\alias{table_11_8}
\title{Table 11.8 Median Salaries of Full Professors in Statistics, 2007}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
5 variables:
\describe{
\item{years_rank}{}
\item{year}{}
\item{year2:}{midpoint of the range}
\item{count}{}
\item{median:}{average salary}
}
}
\source{
American Statistical Association, “2007 Salary Report.”
}
\usage{
table_11_8
}
\description{
Median Salaries of Full Professors in Statistics, 2007
}
\keyword{datasets}
