% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_9.R
\docType{data}
\name{table_11_9}
\alias{table_11_9}
\title{Table 11.9 Stock and Consumer Prices, Post–World War II Period}
\format{
A \link[tibble:tibble-package]{tibble} with 20 rows and
4 variables:
\describe{
\item{obs:}{observation}
\item{country}{}
\item{y:}{stock prices; rate of change, \% per year}
\item{x:}{consumer prices; rate of change, \% per year}
}
}
\source{
Phillip Cagan, Common Stock Values and Inﬂation:
The Historical Record of Many Countries, National Bureau of Economic Research,
Suppl., March 1974, Table 1, p. 4.
}
\usage{
table_11_9
}
\description{
Stock and Consumer Prices, Post–World War II Period (through 1969)
}
\keyword{datasets}
