% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_12_4.R
\docType{data}
\name{table_12_4}
\alias{table_12_4}
\title{Table 12.4 Indexes of Real Compensation and Productivity, U.S., 1960–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 46 rows and
3 variables:
\describe{
\item{year}{}
\item{y:}{index of real compensation per hour, business sector (1992 = 100)}
\item{x:}{index of output, business sector (1992 = 100)}
}
}
\source{
Economic Report of the President, 2007, Table B-49.
}
\usage{
table_12_4
}
\description{
Indexes of Real Compensation and Productivity, U.S., 1960–2005
(Index numbers, 1992 = 100; quarterly data seasonally adjusted)
}
\keyword{datasets}
