% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_12_8.R
\docType{data}
\name{table_12_8}
\alias{table_12_8}
\title{Table 12.8 Smoking and Lung Cancer}
\format{
A \link[tibble:tibble-package]{tibble} with 15 rows and
4 variables:
\itemize{
\item \strong{y:} personal consumption expenditure, $, billions of 1958
\item \strong{x:} time; 1 = 1956, ..., 15 = 1970
\item \strong{y_hat:} estimated y
\item \strong{u:} residuals
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 462. Data for Y_hat obtained from the regression Yt = β0 + β1Xt + ut.
}
\usage{
table_12_8
}
\description{
Smoking and Lung Cancer
}
\keyword{datasets}
