% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_12_9.R
\docType{data}
\name{table_12_9}
\alias{table_12_9}
\title{Table 12.9 Inventories and Sales in U.S. Manufacturing, 1950–1991}
\format{
A \link[tibble:tibble-package]{tibble} with 41 rows and
4 variables:
\describe{
\item{year}{}
\item{sales:}{annual data are averages of monthly, not seasonally adjusted, ﬁgures}
\item{inventories:}{seasonally adjusted, end of period ﬁgures beginning 1982 are not comparable with earlier period}
\item{ratio:}{inventories/sales}
}
}
\source{
Economic Report of the President, 1993, Table B-53, p. 408.
}
\usage{
table_12_9
}
\description{
Inventories and Sales in U.S. Manufacturing, 1950–1991 (millions of dollars)
}
\keyword{datasets}
