% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_13_3.R
\docType{data}
\name{table_13_3}
\alias{table_13_3}
\title{Table 13.3 Per Capita Personal Consumption Expenditure and per Capita Personal Disposable Income , U.S.}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
3 variables:
\describe{
\item{year}{}
\item{ppce:}{personal consumption expenditure}
\item{pdpi:}{personal disposable income}
}
}
\source{
Economic Report of the President, 2007.
}
\usage{
table_13_3
}
\description{
Per Capita Personal Consumption Expenditure (PPCE) and per Capita Personal Disposable Income (PDPI), U.S., 1970–2005
}
\keyword{datasets}
