% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_14_1.R
\docType{data}
\name{table_14_1}
\alias{table_14_1}
\title{Table 14.1 Advisory Fees Charged and Asset Size}
\format{
A \link[tibble:tibble-package]{tibble} with 12 rows and
3 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{y:} fee, \%
\item \strong{x:} net asset value, billions of dollars
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 528.
}
\usage{
table_14_1
}
\description{
Advisory Fees Charged and Asset Size
}
\keyword{datasets}
