% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_14_2.R
\docType{data}
\name{table_14_2}
\alias{table_14_2}
\title{Table 14.2 U.S. Population}
\format{
A \link[tibble:tibble-package]{tibble} with 38 rows and
2 variables:
\describe{
\item{year}{}
\item{population}{}
}
}
\source{
Economic Report of the President, 2008.
}
\usage{
table_14_2
}
\description{
U.S. Population (Millions)
}
\keyword{datasets}
