% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_14_3.R
\docType{data}
\name{table_14_3}
\alias{table_14_3}
\title{Table 14.3 Production Function Data for the Mexican Economy}
\format{
A \link[tibble:tibble-package]{tibble} with 20 rows and
4 variables:
\itemize{
\item \strong{obs:} observation; year
\item \strong{gdp:} GDP is in millions of 1960 pesos
\item \strong{labor:} labor is in thousands of people
\item \strong{capital:} capital is in millions of 1960 pesos
}
}
\source{
Victor J. Elias, Sources of Growth: A Study of Seven Latin American Economies,
International Center for Economic Growth, ICS Press, San Francisco, 1992, Tables E-5, E-12, E-14.
}
\usage{
table_14_3
}
\description{
Production Function Data for the Mexican Economy
}
\keyword{datasets}
