% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_14_4.R
\docType{data}
\name{table_14_4}
\alias{table_14_4}
\title{Table 14.4}
\format{
A \link[tibble:tibble-package]{tibble} with 12 rows and
3 variables:
\itemize{
\item \strong{occupation}
\item \strong{smoking}
\item \strong{cancer}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 540.
}
\usage{
table_14_4
}
\description{
Table 14.4
}
\keyword{datasets}
