% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_1.R
\docType{data}
\name{table_15_1}
\alias{table_15_1}
\title{Table 15.1 Hypothetical Data on Home Ownership and Income}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
3 variables:
\itemize{
\item \strong{family}
\item \strong{y:} home ownership; 1 = owns a house, 0 = does not own a house
\item \strong{x:} family income, thousands of $
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 547.
}
\usage{
table_15_1
}
\description{
Data on home ownership, Y, and family income, X, for 40 families.
}
\keyword{datasets}
