% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_10.R
\docType{data}
\name{table_15_10}
\alias{table_15_10}
\title{Table 15.10 Estimating the Index Ii from the Standard Normal CDF}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
2 variables:
\itemize{
\item \strong{p_hat}
\item \strong{i:} I = F^(-1) * p_hat
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 568.
}
\usage{
table_15_10
}
\description{
Estimating the Index Ii from the Standard Normal CDF
}
\keyword{datasets}
