% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_2.R
\docType{data}
\name{table_15_2}
\alias{table_15_2}
\title{Table 15.2 Actual Y, Estimated Y, and Weights wi for the Home Ownership Example}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
4 variables:
\describe{
\item{y:}{actual y}
\item{y_hat:}{estimated y}
\item{w_hat:}{weights}
\item{sqrt_w:}{square root weights}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 548.
}
\usage{
table_15_2
}
\description{
Actual Y, Estimated Y, and Weights wi for the Home Ownership Example
}
\keyword{datasets}
