% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_22.R
\docType{data}
\name{table_15_22}
\alias{table_15_22}
\title{Table 15.22 Toxicity Study and Rotenone on Chrysanthemum Aphis}
\format{
A \link[tibble:tibble-package]{tibble} with 5 rows and
5 variables:
\describe{
\item{x:}{concentration, milligrams per liter}
\item{log_x:}{concentration, milligrams per liter}
\item{Ni:}{total}
\item{ni:}{death}
\item{p_hat:}{ni/Ni}
}
}
\source{
D. J. Fennet, Probit Analysis, Cambridge University Press, London, 1964.
}
\usage{
table_15_22
}
\description{
Data on the results of spraying rotenone of different concentrations on the
chrysanthemum aphis in batches of approximately fifty.
}
\keyword{datasets}
