% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_24.R
\docType{data}
\name{table_15_24}
\alias{table_15_24}
\title{Table 15.24 Price of Soda with Discount Coupon}
\format{
A \link[tibble:tibble-package]{tibble} with 11 rows and
3 variables:
\describe{
\item{X:}{price discount, ¢}
\item{Ni:}{sample size}
\item{ni:}{number of coupons redeemed}
}
}
\source{
Douglas C. Montgomery and Elizabeth A. Peck, Introduction to Linear Regression
Analysis, John Wiley & Sons, New York, 1982, p. 243 (notation changed).
}
\usage{
table_15_24
}
\description{
Price of Soda with Discount Coupon
}
\keyword{datasets}
