% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_26.R
\docType{data}
\name{table_15_26}
\alias{table_15_26}
\title{Table 15.26 Hypothetical Data Set Generated by the Model Y = -1 + 3X + e}
\format{
A \link[tibble:tibble-package]{tibble} with 35 rows and
3 variables:
\describe{
\item{y:}{y = -1 + 3x + e}
\item{y_star:}{y_star = 1 se y > 0; 0 otherwise}
\item{x}{}
}
}
\source{
William E. Becker and Donald M. Waldman, “A Graphical Interpretation
of Probit Coefﬁcients,” Journal of Economic Education, Fall 1989, Table 1, p. 373.
}
\usage{
table_15_26
}
\description{
Hypothetical Data Set Generated by the Model Y = -1 + 3X + e
}
\keyword{datasets}
