% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_27.R
\docType{data}
\name{table_15_27}
\alias{table_15_27}
\title{Table 15.27 Data for 2,000 women regarding work}
\format{
A \link[tibble:tibble-package]{tibble} with 2000 rows and
15 variables:
\describe{
\item{c1}{}
\item{c2}{}
\item{u}{}
\item{v}{}
\item{county}{}
\item{age}{}
\item{education:}{number of years of schooling}
\item{married:}{married = 1, married; married = 0, otherwise}
\item{children:}{number of children}
\item{select}{}
\item{wagefull}{}
\item{wage}{}
\item{lw}{}
\item{work}{work = 1, woman works; work = 0, otherwise}
\item{lwf}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 589.
}
\usage{
table_15_27
}
\description{
Data for 2,000 women regarding work
}
\keyword{datasets}
