% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_28.R
\docType{data}
\name{table_15_28}
\alias{table_15_28}
\title{Table 15.28  The product of education and income has any effect on the probability of becoming a smoker}
\format{
A \link[tibble:tibble-package]{tibble} with 1196 rows and
7 variables:
\describe{
\item{age}{}
\item{adeeduc}{}
\item{educ}{}
\item{income}{}
\item{pcigs79}{}
\item{smoker}{}
\item{educ*income}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 589.
}
\usage{
table_15_28
}
\description{
The product of education and income has any effect on the probability of becoming a smoker.
}
\keyword{datasets}
