% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_29.R
\docType{data}
\name{table_15_29}
\alias{table_15_29}
\title{Table 15.29}
\format{
A \link[tibble:tibble-package]{tibble} with 178 rows and
12 variables:
\itemize{
\item \strong{cancer:} = 1, had breast cancer; = 0, did not have breast cancer
\item \strong{age:} age of subject
\item \strong{higd:} highest grade completed in school
\item \strong{degree}
\item \strong{chk:} = 0 if subject did not undergo regular medical checkups; = 1 if subject did undergo regular checkups
\item \strong{agpi:}age at first pregnancy
\item \strong{men_age}
\item \strong{miscarriages:} number of miscarriages
\item \strong{births}
\item \strong{weight:} weight of subject
\item \strong{age_period}
\item \strong{marital_status}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 589.
}
\usage{
table_15_29
}
\description{
Predicting whether a woman will contract breast cancer or not.
}
\keyword{datasets}
