% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_4.R
\docType{data}
\name{table_15_4}
\alias{table_15_4}
\title{Table 15.4 Hypothetical Data on Income, Number of Families at Income Xi and Number of Families Owning a House}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
3 variables:
\describe{
\item{X:}{income, thousands of $}
\item{N:}{number of families at income, Xi}
\item{n:}{number of families owning a house}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 556.
}
\usage{
table_15_4
}
\description{
data on several families grouped or replicated (repeat observations) according to
income level and the number of families owning a house at each income level
}
\keyword{datasets}
