% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_16_1.R
\docType{data}
\name{table_16_1}
\alias{table_16_1}
\title{Table 16.1 Costs of six airline firms for the period 1970-1984}
\format{
A \link[tibble:tibble-package]{tibble} with 90 rows and
10 variables:
\itemize{
\item \strong{i:} airline id
\item \strong{t:} year id
\item \strong{c:} total cost, in $1000
\item \strong{q:} output, in revenue passenger miles
\item \strong{pf:} fuel price
\item \strong{lf:} load factor
\item \strong{ln_c:} natural logarithm (ln) of total cost
\item \strong{ln_q:} natural logarithm (ln) of output
\item \strong{ln_pf:} natural logarithm (ln) of fuel price
\item \strong{ln_lf:} natural logarithm (ln) of load factor
}
}
\source{
William H. Greene, Econometric Analysis, 6th ed., 2008.
Data are located at \url{http://pages.stern.nyu.edu/~wgreen/Text/econometricanalysis.htm}
}
\usage{
table_16_1
}
\description{
The data analyzes the costs of six airline ﬁrms for the period 1970–1984,
for a total of 90 panel data observations.
}
\keyword{datasets}
