% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_16_16.R
\docType{data}
\name{table_16_16}
\alias{table_16_16}
\title{Table 16.16 Unemployment Rate and Hourly Compensation in Manufacturing, in the United States, Canada, and the United Kingdom, 1980–2006.}
\format{
A \link[tibble:tibble-package]{tibble} with 27 rows and
7 variables:
\describe{
\item{year}{}
\item{comp_us:}{hourly compensation in manufacturing in United States, U.S. dollars, 1992-100}
\item{un_us:}{unemployment rate in United States, \%}
\item{comp_can:}{hourly compensation in manufacturing in Canada, U.S. dollars, 1992-100}
\item{un_can:}{unemployment rate in Canada, \%}
\item{comp_uk:}{hourly compensation in manufacturing in United Kingdom, U.S. dollars, 1992-100}
\item{un_uk:}{unemployment rate in United Kingdom, \%}
}
}
\source{
Economic Report of the President, January 2008, Table B-109.
}
\usage{
table_16_16
}
\description{
Unemployment Rate and Hourly Compensation in Manufacturing, in the United States, Canada, and the United Kingdom, 1980–2006.
}
\keyword{datasets}
