% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_16_17.R
\docType{data}
\name{table_16_17}
\alias{table_16_17}
\title{Table 16.17 Gasoline demand in the 18 OECD countries}
\format{
A \link[tibble:tibble-package]{tibble} with 342 rows and
7 variables:
\itemize{
\item \strong{country}
\item \strong{ctry_code:} country code, in all 18 OECD countries
\item \strong{year:} annual observations from 1960–1978
\item \strong{lgaspcar:} gasoline consumption per car
\item \strong{lincomep:} real income per capita
\item \strong{lrpmg:} real gasoline price
\item \strong{lcarpcap:} number of cars per capita
}
}
\source{
B. H. Baltagi and J. M. Grifﬁn, “Gasoline Demand in the OECD:
An Application of Pooling and Testing Procedures,” European Economic Review,
vol. 22, 1983, pp. 117–137. The data for 18 OECD countries for the years 1960–1978
can be obtained from: \url{http://www.wiley.com/legacy/wileychi/baltagi/supp/Gasoline.dat},
or from the textbook website, Table 16.17.
}
\usage{
table_16_17
}
\description{
Gasoline demand in the 18 OECD countries. Values in table are logged already.
}
\keyword{datasets}
