% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_16_18.R
\docType{data}
\name{table_16_18}
\alias{table_16_18}
\title{Table 16.18 Responsiveness of aid to recipient countries’ economic and physical needs, civil/political rights, and government effectiveness}
\format{
A \link[tibble:tibble-package]{tibble} with 395 rows and
24 variables:
\describe{
\item{countryid}{}
\item{country}{}
\item{cntry_code}{}
\item{year}{}
\item{aidcap}{}
\item{raidcap}{}
\item{infmort}{}
\item{infmort_2}{}
\item{pop}{}
\item{raid}{}
\item{civlib}{}
\item{polright}{}
\item{civil_polrts}{}
\item{goveff}{}
\item{regqual}{}
\item{rgdpcap1}{}
\item{rgdpcap}{}
\item{rgdcap_2}{}
\item{popmil}{}
\item{popmil_2}{}
\item{deflator}{}
\item{y1}{}
\item{y2}{}
\item{y3}{}
}
}
\source{
Subhayu Bandyopadhyay and Howard J. Wall, “The Determinants of Aid in
the Post-Cold War Era,” Review,  Federal Reserve Bank of St. Louis,
November/December 2007, vol. 89, number 6, pp. 533–547.
The article and data can be found at: \url{http://research.stlouisfed.org/publications/review/past/2007}
in the November/December, Vol. 89, No. 10 section.
}
\usage{
table_16_18
}
\description{
Data to estimate the responsiveness of aid to recipient countries’ economic and physical needs,
civil/political rights, and government effectiveness. The data are for 135 countries for three years.
}
\keyword{datasets}
