% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_10.R
\docType{data}
\name{table_17_10}
\alias{table_17_10}
\title{Table 17.10 Investment in FixedPlant and Equipment in Manufacturing and Manufacturing Sales, United States, 1970-1991}
\format{
A \link[tibble:tibble-package]{tibble} with 22 rows and
3 variables:
\describe{
\item{year}{}
\item{y:}{plant expenditure, seasonally adjusted, billions of $}
\item{x2:}{manufacturing sales, seasonally adjusted, billions of $}
}
}
\source{
Economic Report of the President, 1993. Data on Y from Table B-52, p. 407;
data on X2 from Table 8-53, p. 408.
}
\usage{
table_17_10
}
\description{
Investment in Fixed Plant and Equipment in Manufacturing Y and Manufacturing Sales X2 in Billions
of Dollars, Seasonally Adjusted, United States, 1970–1991
}
\keyword{datasets}
