% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_11.R
\docType{data}
\name{table_17_11}
\alias{table_17_11}
\title{Table 17.11 Investments, Sales, and Interest Rate, United States, 1960–1999}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{investment:} private ﬁxed investment in information processing equipment and software, billions of dollars, seasonally adjusted
\item \strong{sales:} sales in total manufacturing and trade, millions of dollars, seasonally adjusted
\item \strong{interest:} Moody’s Aaa corporate bond rate, \%
}
}
\source{
Economic Report of the President, 2001, Tables B-18, B-57, and B-73.
}
\usage{
table_17_11
}
\description{
Data on private ﬁxed investment in information processing and equipment #' (Y,in billions of dollars),
sales in total manufacturing and trade (X2, in millions of dollars), and interest rate
(X3, Moody’s Aaa corporate bond rate, percent).
}
\keyword{datasets}
