% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_12.R
\docType{data}
\name{table_17_12}
\alias{table_17_12}
\title{Table 17.12 Compensation, Productivity and Unemployment Rate, United States, 1960–1999}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{compens:} index of real compensation per hour (1992 = 100)
\item \strong{produc:} index of output per hour (1992 = 100)
\item \strong{unrate:} civilian unemployment rate, \%
}
}
\source{
Economic Report of the President, 2001, Table B-49, p. 332.
}
\usage{
table_17_12
}
\description{
Data on indexes of real compensation per hour (Y) and output per hour (X2),
with both indexes to base 1992 = 100, in the business sector of the U.S. economy
for the period 1960–1999, as well as the civilian unemployment rate (X3) for the same period.
}
\keyword{datasets}
