% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_13.R
\docType{data}
\name{table_17_13}
\alias{table_17_13}
\title{Table 17.13 Macroeconomic Data for the Greek Economy, 1960–1995}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
6 variables:
\describe{
\item{year}{}
\item{pc:}{real desired private consumption expenditure}
\item{pdi:}{real private disposable income}
\item{grossinv}{}
\item{gnp}{}
\item{lti:}{long-term interest rate}
}
}
\source{
H. R. Seddighi, K. A. Lawler, and A. V. Katos, Econometrics:
A Practical Approach, Routledge, London, 2000, p. 158.
}
\usage{
table_17_13
}
\description{
Macroeconomic Data for the Greek Economy, 1960–1995
}
\keyword{datasets}
