% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_2.R
\docType{data}
\name{table_17_2}
\alias{table_17_2}
\title{Table 17.2 PPCE and PPDI, 1959–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 48 rows and
3 variables:
\itemize{
\item \strong{year}
\item \strong{ppce:} per capita personal consumption expenditure in chained 2000 dollars
\item \strong{ppdi:} per capita personal disposable income in chained 2000 dollars
}
}
\source{
Economic Report of the President, 2007, Table B-31
}
\usage{
table_17_2
}
\description{
PPCE and PPDI, 1959–2006
}
\keyword{datasets}
