% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_5.R
\docType{data}
\name{table_17_5}
\alias{table_17_5}
\title{Table 17.5 Money, Interest Rate, Price Index, and GDP, Canada}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
5 variables:
\itemize{
\item \strong{obs:} observation; year and quarter
\item \strong{m1:} C$, millions
\item \strong{r:} 90-day prime corporate interest rate, \%
\item \strong{p:} implicit price deflator (1981 = 100)
\item \strong{gdp:} C$, millions (1981 prices)
}
}
\source{
Rao, op. cit., pp. 210–213.
}
\usage{
table_17_5
}
\description{
Money, Interest Rate, Price Index, and GDP, Canada
}
\keyword{datasets}
