% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_7.R
\docType{data}
\name{table_17_7}
\alias{table_17_7}
\title{Table 17.7 Private Consumption Expenditure and GDP, Sri Lanka}
\format{
A \link[tibble:tibble-package]{tibble} with 27 rows and
3 variables:
\describe{
\item{year}{}
\item{pcon:}{private consumption expenditure}
\item{gdp:}{gross domestic product}
}
}
\source{
Chandan Mukherjee, Howard White, and Marc Wuyts. Econometrics and Data Analysis for
Developing Countries, Routledge, New York, 1998. The original data are from World Bank’s World Tables.
}
\usage{
table_17_7
}
\description{
Private Consumption Expenditure and GDP, Sri Lanka for the period 1967-1993
}
\keyword{datasets}
