% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_8.R
\docType{data}
\name{table_17_8}
\alias{table_17_8}
\title{Table 17.8 Inventories Y and Sales X,U.S. Manufacturing, and Constructed Z’s}
\format{
A \link[tibble:tibble-package]{tibble} with 46 rows and
3 variables:
\itemize{
\item \strong{year}
\item \strong{inventory:} in millions of dollars, seasonally adjusted
\item \strong{sales:} in millions of dollars, seasonally adjusted
}
}
\source{
Economic Report of the President, 2001, Table B-57, p. 340.
}
\usage{
table_17_8
}
\description{
Inventories Y and Sales X,U.S. Manufacturing, and Constructed Z’s.
}
\keyword{datasets}
