% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_18_2.R
\docType{data}
\name{table_18_2}
\alias{table_18_2}
\title{Table 18.2 Personal Consumption Expenditure, Gross Private Domestic Investment, and GDP, United States, 1970–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 37 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{c:} personal consumption expenditure
\item \strong{i:} gross private domestic investment
\item \strong{y:} gross domestic product
}
}
\source{
Economic Report of the President, 2008, Table B-2.
}
\usage{
table_18_2
}
\description{
Personal Consumption Expenditure, Gross Private Domestic Investment, and GDP, United States, 1970–2006 (billions of 1996 dollars)
}
\keyword{datasets}
