% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_18_3.R
\docType{data}
\name{table_18_3}
\alias{table_18_3}
\title{Table 18.3 Supply and demand for gas}
\format{
A \link[tibble:tibble-package]{tibble} with 296 rows and
6 variables:
\describe{
\item{obs:}{observation}
\item{year_quarter}{}
\item{carsales:}{millions of cars per year}
\item{persincome:}{(personal income, billions of dollars}
\item{pricegas:}{cents per gallon}
\item{quantgas:}{thousands of barrels per day, unleaded}
}
}
\source{
Stephen J. Schmidt, Econometrics, McGraw-Hill, New York, 2005.
See \url{www.mhhe.com/economics}.
}
\usage{
table_18_3
}
\description{
Data on some of the variables that determine demand for and supply of gasoline
in the U.S. from January 1978 to August 2002.
}
\keyword{datasets}
