% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_19_4.R
\docType{data}
\name{table_19_4}
\alias{table_19_4}
\title{Table 19.4 Money, GDP, Interest Rate, and Consumer Price Index, United States, 1970–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 37 rows and
5 variables:
\itemize{
\item \strong{obs:} observation, year
\item \strong{m2:} M2 money supply (billions of dollars)
\item \strong{gdp:} gross domestic product (billions of dollars)
\item \strong{tbrate:} 3-month Treasury bill rate, \%
\item \strong{cpi:} consumer price index (1982–1984 = 100)
}
}
\source{
Economic Report of the President, 2007, Tables B-2, B-60, B-69, B-73.
}
\usage{
table_19_4
}
\description{
Money, GDP, Interest Rate, and Consumer Price Index, United States, 1970–2006
}
\keyword{datasets}
