% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_1.R
\docType{data}
\name{table_1_1}
\alias{table_1_1}
\title{Table 1.1 U.S. Egg Production}
\format{
A \link[tibble:tibble-package]{tibble} with 50 rows and
5 variables:
\itemize{
\item \strong{state}
\item \strong{y1:} eggs produced in 1990 (millions)
\item \strong{y2:} eggs produced in 1991 (millions)
\item \strong{x1:} price per dozen (cents) in 1990
\item \strong{x2:} price per dozen (cents) in 1991
}
}
\source{
World Almanac, 1993, p. 119.
The data are from the Economic Research Service, U.S. Department of Agriculture.
}
\usage{
table_1_1
}
\description{
U.S. Egg Production
}
\keyword{datasets}
