% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_3.R
\docType{data}
\name{table_1_3}
\alias{table_1_3}
\title{Table 1.3 CPI in Seven Industrial Countries, 1980-2005 (1982-1984=100)}
\format{
A \link[tibble:tibble-package]{tibble} with 26 rows and
8 variables:
\describe{
\item{year}{}
\item{usa}{}
\item{canada}{}
\item{japan}{}
\item{france}{}
\item{germany}{}
\item{italy}{}
\item{uk}{}
}
}
\source{
Economic Report of the President, 2007, Table 108, p. 354.
}
\usage{
table_1_3
}
\description{
CPI in Seven Industrial Countries, 1980-2005
}
\keyword{datasets}
