% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_4.R
\docType{data}
\name{table_1_4}
\alias{table_1_4}
\title{Table 1.4 Exchange Rates for Nine Countries: 1985-2006}
\format{
A \link[tibble:tibble-package]{tibble} with 22 rows and
10 variables:
\describe{
\item{year}{}
\item{australia:}{dollar}
\item{canada:}{dollar}
\item{china_p_r:}{yuan}
\item{mexico:}{peso}
\item{south_korea:}{won}
\item{sweden:}{krona}
\item{switzerland:}{franc}
\item{united_kingdom:}{pound}
}
}
\source{
Economic Report of the President, 2007, Table B-110, p. 35.
}
\usage{
table_1_4
}
\description{
Exchange Rates for Nine Countries: 1985-2006
Foreign exchange rates for nine industrialized countries for the years 1985–2006.
Except for the United Kingdom, the exchange rate is deﬁned as the units of foreign currency for one U.S. dollar;
for the United Kingdom, it is deﬁned as the number of U.S. dollars for one U.K. pound.
}
\keyword{datasets}
