% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_5.R
\docType{data}
\name{table_1_5}
\alias{table_1_5}
\title{Table 1.5 Seasonally Adjusted M1 Supply: 1959:01 - 2007:10}
\format{
A \link[tibble:tibble-package]{tibble} with 588 rows and
4 variables:
\itemize{
\item \strong{year_month}
\item \strong{year}
\item \strong{month}
\item \strong{m1:} money stock; billions of dollars
}
}
\source{
Board of Governors, Federal Reserve Bank, USA.
}
\usage{
table_1_5
}
\description{
Seasonally Adjusted M1 Supply: 1959:01 - 2007:10 (billions of dollars)
}
\keyword{datasets}
