% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1_6.R
\docType{data}
\name{table_1_6}
\alias{table_1_6}
\title{Table 1.6 Impact of Advertising Expenditure}
\format{
A \link[tibble:tibble-package]{tibble} with 21 rows and
4 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{firm}
\item \strong{impressions:} in millions
\item \strong{expenditure:} in millions of 1983 dollars
}
}
\source{
\url{http://lib.stat.cmu.edu/DASL/Dataﬁles/tvadsdat.html}
}
\usage{
table_1_6
}
\description{
Impact of Advertising Expenditure
}
\keyword{datasets}
