% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_20_1.R
\docType{data}
\name{table_20_1}
\alias{table_20_1}
\title{Table 20.1 Crop Production, Crop Prices, and per Capita Personal Consumption Expenditures}
\format{
A \link[tibble:tibble-package]{tibble} with 30 rows and
4 variables:
\describe{
\item{obs:}{pbservation; year}
\item{q:}{index of crop production (1996 = 100)}
\item{p:}{index of crop prices received by farmers (1990-1992 = 100)}
\item{x:}{real per capita personal consumption expenditure}
}
}
\source{
Economic Report of the President, 2007.
Data on Q (Table B-99), on P (Table B-101), and on X (Table B-31).
}
\usage{
table_20_1
}
\description{
Crop Production, Crop Prices, and per Capita Personal Consumption Expenditures, 2007 Dollars, United States, 1975–2004
}
\keyword{datasets}
