% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_20_5.R
\docType{data}
\name{table_20_5}
\alias{table_20_5}
\title{Table 20.5 Underlying Data for Klein’s Model I}
\format{
A \link[tibble:tibble-package]{tibble} with 22 rows and
10 variables:
\describe{
\item{year}{}
\item{c:}{consumption expenditure}
\item{p:}{profits}
\item{w:}{private wage bill}
\item{i:}{investment expenditure}
\item{k-1:}{capital stock}
\item{x:}{income after tax}
\item{w':}{government wage bill}
\item{g:}{government expenditure}
\item{t:}{taxes}
}
}
\source{
G. S. Maddala, Econometrics, McGraw-Hill, New York, 1977, p. 238.
}
\usage{
table_20_5
}
\description{
Underlying Data for Klein’s Model I
}
\keyword{datasets}
