% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_21_1.R
\docType{data}
\name{table_21_1}
\alias{table_21_1}
\title{Table 21.1}
\format{
A \link[tibble:tibble-package]{tibble} with 244 rows and
7 variables:
\describe{
\item{year}{}
\item{quarter}{}
\item{dpi:}{real disposable personal income (billions of dollars)}
\item{gdp:}{gross domestic product (billions of dollars)}
\item{pce:}{real personal consumption expenditure (billions of dollars)}
\item{cp:}{corporate proﬁts (billions of dollars)}
\item{dividend:}{ dividends, (billions of dollars)}
}
}
\source{
All the data are collected from FRED, the economic website of the Federal
Reserve Bank of St. Louis. GDP, DPI, and PCE are in constant dollars, here 2000 dollars.
CP and Dividend are in nominal dollars.
}
\usage{
table_21_1
}
\description{
The time period covered is from 1947–I to 2007–IV, for a total of 244 quarters,
and all data are seasonally adjusted at the annual rate
}
\keyword{datasets}
