% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_22_5.R
\docType{data}
\name{table_22_5}
\alias{table_22_5}
\title{Table 22.5 Help wanted Index and the Unemployment rate, USA, January 1969 to January 2000}
\format{
A \link[tibble:tibble-package]{tibble} with 373 rows and
3 variables:
\itemize{
\item \strong{time:} 1 = january 1969; 373 = january 2000
\item \strong{hwi:} help-wanted index
\item \strong{unrate:} unemployment rate (\%)
}
}
\source{
Marc A. Giammatteo (West Point, Class of 2000), “The Relationship between
the Help Wanted Index and the Unemployment Rate,” unpublished term paper.
}
\usage{
table_22_5
}
\description{
Help wanted Index and the Unemployment rate, USA, January 1969 to January 2000
}
\keyword{datasets}
