% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_22_6.R
\docType{data}
\name{table_22_6}
\alias{table_22_6}
\title{Table 22.6 The yen/dollar exchange rate}
\format{
A \link[tibble:tibble-package]{tibble} with 448 rows and
4 variables:
\describe{
\item{yen_exchange:}{}
\item{time:}{}
\item{year}{}
\item{month}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 797.
}
\usage{
table_22_6
}
\description{
The yen/dollar exchange rate
}
\keyword{datasets}
