% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_22_7.R
\docType{data}
\name{table_22_7}
\alias{table_22_7}
\title{Table 22.7 Three-month (TB3M) and six-month (TB6M) Treasury bill rates}
\format{
A \link[tibble:tibble-package]{tibble} with 324 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{month}
\item \strong{tb3m:} three-month treasury bill rate
\item \strong{tb6m:} six-month treasury bill rate
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 800.
}
\usage{
table_22_7
}
\description{
Data on three-month (TB3M) and six-month (TB6M) Treasury bill rates from
January 1, 1982, to March 2008, for a total of 315 monthly observations.
}
\keyword{datasets}
