% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_10.R
\docType{data}
\name{table_2_10}
\alias{table_2_10}
\title{Table 2.10 SAT Reasoning Test Classified by Family Income}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
9 variables:
\describe{
\item{family_income}{}
\item{middle_income:}{middle income in the interval}
\item{n_test_takers:}{}
\item{cr_mean}{critical reading mean}
\item{cr_sd}{critical reading standart deviation}
\item{m_mean}{mathematics mean}
\item{m_sd}{mathematics standart deviation}
\item{w_mean}{writing mean}
\item{w_sd}{writing standart deviation}
}
}
\source{
College Board, 2007. College-Bound Seniors, Table 11.
}
\usage{
table_2_10
}
\description{
SAT Reasoning Test Classified by Family Income
}
\keyword{datasets}
