% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_5.R
\docType{data}
\name{table_2_5}
\alias{table_2_5}
\title{Table 2.5 Another Random Sample from the Population of Table 2.1}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
2 variables:
\itemize{
\item \strong{y}
\item \strong{x}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 43.
}
\usage{
table_2_5
}
\description{
Another Random Sample from the Population of Table 2.1
}
\keyword{datasets}
