% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_7.R
\docType{data}
\name{table_2_7}
\alias{table_2_7}
\title{Table 2.7 Labor Force Participation Data for U.S. for 1980-2006}
\format{
A \link[tibble:tibble-package]{tibble} with 27 rows and
7 variables:
\describe{
\item{year}{}
\item{clfprm:}{civilian labor force participation rate, male (\%), Table B-39, p. 277.}
\item{clfprf:}{civilian labor force participation rate, female (\%), Table B-39, p. 277.}
\item{unrm:}{civilian unemployment rate, male (\%), Table B-42, p. 280.}
\item{unrf:}{civilian unemployment rate, male (\%), Table B-42, p. 280.}
\item{ahe82:}{average hourly earnings (1982 dollars), Table B-47, p.286.}
\item{ahe:}{average hourly earnings (current dollars), Table B-47, p.286.}
}
}
\source{
Economic Report of the President, 2007.
}
\usage{
table_2_7
}
\description{
Labor Force Participation Data for U.S. for 1980-2006
}
\keyword{datasets}
