% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_8.R
\docType{data}
\name{table_2_8}
\alias{table_2_8}
\title{Table 2.8 Food and Total Expenditure (Rupees)}
\format{
A \link[tibble:tibble-package]{tibble} with 55 rows and
3 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{foodexp:} food expenditure
\item \strong{totalexp:} total expenditure
}
}
\source{
Chandan Mukherjee, Howard White, and Marc Wuyts, Econometrics and
Data Analysis for Developing Countries, Routledge, New York, 1998, p. 457.
}
\usage{
table_2_8
}
\description{
Food and Total Expenditure (Rupees)
}
\keyword{datasets}
