% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_9.R
\docType{data}
\name{table_2_9}
\alias{table_2_9}
\title{Table 2.9 Total Group Mean SAT Reasoning Test Scores: College-Bound Seniors, 1972-2007}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
7 variables:
\itemize{
\item \strong{year}
\item \strong{cr_male:} male in critical reading
\item \strong{cr_female:} female in critical reading
\item \strong{cr_total:} total in critical reading
\item \strong{m_male:} male in mathematics
\item \strong{m_female:} female in mathematics
\item \strong{m_total:} total in mathematics
}
}
\source{
College Board, 2007.
}
\usage{
table_2_9
}
\description{
Total Group Mean SAT Reasoning Test Scores: College-Bound Seniors, 1972-2007
}
\note{
For 1972–1986 a formula was applied to the original mean and standard deviation
to convert the mean to the recentered scale. For 1987–1995 individual student scores
were converted to the recentered scale and then the mean was recomputed.
From 1996–1999, nearly all students received scores on the recentered scale.
Any score on the original scale was converted to the recentered scale prior to
computing the mean. From 2000–2007, all scores are reported on the recentered scale.
}
\keyword{datasets}
