% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_2.R
\docType{data}
\name{table_3_2}
\alias{table_3_2}
\title{Table 3.2 Raw Data Based on Table 2.6}
\format{
A \link[tibble:tibble-package]{tibble} with 13 rows and
3 variables:
\describe{
\item{obs:}{observation}
\item{y:}{mean hourly wage}
\item{x}{years of schooling}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 79.
}
\usage{
table_3_2
}
\description{
Raw Data Based on Table 2.6
}
\keyword{datasets}
