% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_3.R
\docType{data}
\name{table_3_3}
\alias{table_3_3}
\title{Table 3.3 Cellphone Subscribers x Personal Computers x Per Capita Income in Selected Countries, 2003.}
\format{
A \link[tibble:tibble-package]{tibble} with 34 rows and
7 variables:
\describe{
\item{country}{}
\item{cellphone}{}
\item{pcs}{}
\item{pcapincome}{per capita income ($)}
\item{ln_cellphone:}{natural log (ln) of the variable cellphone}
\item{ln_pc:}{natural log (ln) of the variable pcs}
\item{ln_income:}{natural log (ln) of the variable pcapincome}
}
}
\source{
Statistical Abstract of the United States, 2006, Table 1364 for data on
cellphones and computers and Table 1327 for purchasing-power adjusted per
capita income.
}
\usage{
table_3_3
}
\description{
Number of Cellular Phone Subscribers per Hundred Persons and Number of Personal
Computers per 100 Persons and Per Capita Income in Selected Countries for 2003.
}
\details{
Note: The data on cell phones and personal computers are per 100 persons.
}
\keyword{datasets}
