% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_5.R
\docType{data}
\name{table_3_5}
\alias{table_3_5}
\title{Table 3.5 Rank of 10 students in midterm and final examinations in statistics}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
3 variables:
\describe{
\item{student}{}
\item{midterm}{}
\item{final}{}
}
}
\source{
Gujarati, Damodar N., and Dawn C. Porter. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 87.
}
\usage{
table_3_5
}
\description{
Rank of 10 students in midterm and final examinations in statistics
}
\keyword{datasets}
