% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_6.R
\docType{data}
\name{table_3_6}
\alias{table_3_6}
\title{Table 3.6 Productivity and Related Data, Business Sector 1960–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 46 rows and
5 variables:
\itemize{
\item \strong{year}
\item \strong{outbus:} output per hour of all persons on business sector
\item \strong{outnbus:} output per Hour of all persons on nonfarm business sector
\item \strong{rcombus:} real compensation per hour on business sector
\item \strong{rcomnbus:} real compensation per hour on nonfarm business sector
}
}
\usage{
table_3_6
}
\description{
Productivity and Related Data, Business Sector 1960–2005
(Index numbers, 1992 = 100; quarterly data seasonally adjusted)
}
\note{
Output refers to real gross domestic product in the sector;
Wages and salaries of employees plus employers’ contributions for social insurance and private beneﬁt plans;
Hourly compensation divided by the consumer price index for all urban consumers for recent quarters.

#' @source Economic Report of the President, 2007, Table 49.
}
\keyword{datasets}
