% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_7.R
\docType{data}
\name{table_3_7}
\alias{table_3_7}
\title{Table 3.7 Gold Prices, New York Stock Exchange Index, and Consumer Price Index for U.S. for 1974–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 33 rows and
4 variables:
\describe{
\item{year}{}
\item{gold_price: }{Gold Prices}
\item{nyse: }{New York Stock Exchange Index}
\item{cpi: }{Consumer Price Index}
}
}
\source{
Gujarati, Damodar N., and Dawn C. Porter. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 90.
}
\usage{
table_3_7
}
\description{
Gold Prices, New York Stock Exchange Index, and Consumer Price Index for U.S. for 1974–2006
}
\keyword{datasets}
