% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_8.R
\docType{data}
\name{table_3_8}
\alias{table_3_8}
\title{Table 3.8 Nominal and Real Gross Domestic Product, 1959–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 47 rows and
3 variables:
\describe{
\item{year}{}
\item{ngdp:}{nominal gross domestic product; in billions of dollars}
\item{rgdp:}{real gross domestic product; in billions of chained 2000 dollars}
}
}
\source{
Source: Economic Report of the President, 2007. Table B-1 and B-2.
}
\usage{
table_3_8
}
\description{
Nominal and Real Gross Domestic Product, 1959–2005
(billions of dollars, except as noted; quarterly data at seasonally adjusted
annual rates; RGDP in billions of chained 2000 dollars)
}
\keyword{datasets}
