% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_5_10.R
\docType{data}
\name{table_5_10}
\alias{table_5_10}
\title{Table 5.10 CPI and PPI, USA, 1980–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 27 rows and
3 variables:

\describe{
\item{year}{}
\item{cpi:}{consumer price index}
\item{ppi:}{producer price index, also called wholesale price index (WPI); total finished goods}
}
}
\source{
Economic Report of the President, 2007, Tables B-62 and B-65.
}
\usage{
table_5_10
}
\description{
CPI and PPI, USA, 1980–2006
}
\keyword{datasets}
