% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_5_11.R
\docType{data}
\name{table_5_11}
\alias{table_5_11}
\title{Table 5.11 Smoking and Lung Cancer}
\format{
A \link[tibble:tibble-package]{tibble} with 25 rows and
3 variables:
\describe{
\item{occupation}{}
\item{smoking}{}
\item{cancer}{}
}
}
\source{
\url{http://lib.stat.cmu.edu/DASL/Dataﬁles/SmokingandCancer.html}
}
\usage{
table_5_11
}
\description{
Smoking and Lung Cancer
}
\keyword{datasets}
