% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_5_5.R
\docType{data}
\name{table_5_5}
\alias{table_5_5}
\title{Table 5.5 Average Salary and Per Pupil Spending (dollars), 1985}
\format{
A \link[tibble:tibble-package]{tibble} with 51 rows and
3 variables:
\describe{
\item{obs:}{observation}
\item{salary}{}
\item{spending}{}
}
}
\source{
National Education Association, as reported by Albuquerque Tribune, Nov. 7, 1986.
}
\usage{
table_5_5
}
\description{
Average Salary and Per Pupil Spending (dollars), 1985
}
\keyword{datasets}
