% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_5_6.R
\docType{data}
\name{table_5_6}
\alias{table_5_6}
\title{Table 5.6 GNP and Four Measures of Money Stock}
\format{
A \link[tibble:tibble-package]{tibble} with 14 rows and
6 variables:
\itemize{
\item \strong{year}
\item \strong{gnp:} gross national product ($ billion)
\item \strong{m1:} currency + demand deposits + travelers checks and other checkable deposits (OCDs)
\item \strong{m2:} M1 + overnight RPs and Eurodollars + MMMF (Money market mutual fund) balances + MMDAs (Money market deposit accounts) + savings and small deposits
\item \strong{m3:} M2 + large time deposits + term RPs + institutional MMMF
\item \strong{l:} M3 + other liquid assets
}
}
\source{
Economic Report of the President, 1985, GNP data from Table B-1, p. 232;
money stock data from Table B-61, p. 303.
}
\usage{
table_5_6
}
\description{
GNP and Four Measures of Money Stock (1970-1983)
}
\keyword{datasets}
