% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_5_9.R
\docType{data}
\name{table_5_9}
\alias{table_5_9}
\title{Table 5.9 The Hamburger Standard}
\format{
A \link[tibble:tibble-package]{tibble} with 47 rows and
6 variables:
\describe{
\item{country}{}
\item{bmcplc:}{Big Mac Prices In Local Currency}
\item{bmcpd:}{Big Mac Prices In Dollars}
\item{ippp:}{implied PPP of the Dollar (PPP is Purchasing Power Parity)}
\item{ader:}{actual Dollar exchange rate, Jan 31st}
\item{u_o_vag:}{under (-) / over (+) valuation against the Dollar (\%)}
}
}
\source{
McDonald’s; The Economist, February 1, 2007.
}
\usage{
table_5_9
}
\description{
The Hamburger Standard
}
\keyword{datasets}
