% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_1.R
\docType{data}
\name{table_6_1}
\alias{table_6_1}
\title{Table 6.1 Excess Returns (Yt, \%) (Xt, \%)}
\format{
A \link[tibble:tibble-package]{tibble} with 240 rows and
3 variables:
\itemize{
\item \strong{obs:} year and month
\item \strong{Y:} excess returns (Yt, \%)
\item \strong{X:} excess returns (Xt, \%)
}
}
\source{
These data, originally obtained from DataStream databank,
are reproduced from Christiaan Heij et al., Econometrics Methods with Applications
in Business and Economics, Oxford University Press, Oxford, U.K., 2004.
}
\usage{
table_6_1
}
\description{
Data on excess returns Yt (\%) on an index of 104 stocks in the sector of
cyclical consumer goods and excess returns Xt (\%) on the overall stock market
index for the U.K. for the monthly data for the period 1980–1999,
for a total of 240 observations. Excess return refers to return in excess of
return on a riskless asset.
}
\keyword{datasets}
