% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_10.R
\docType{data}
\name{table_6_10}
\alias{table_6_10}
\title{Table 6.10 Advertising Expenditure and Total Expenditure (in £ millions) for 29 Product Categories in the U.K.}
\format{
A \link[tibble:tibble-package]{tibble} with 29 rows and
9 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{product:} product category
\item \strong{adexp:} advertising expenditure (£, millions)
\item \strong{conexp:} total consumer expenditure (£, millions)
\item \strong{ratio:} (adexp)/(conexp*1000)
\item \strong{lnadexp:} natural log (ln) of advertising expenditure
\item \strong{lnconexp:} natural log (ln) of total consumer expenditure
\item \strong{1/adexp:}
\item \strong{1/conexp:}
}
}
\source{
Advertising Statistics Yearbook 1996 \url{http://www.economicswebinstitute.org/ecdata.htm}
}
\usage{
table_6_10
}
\description{
Advertising Expenditure and Total Expenditure (in £ millions) for 29 Product Categories in the U.K.
}
\keyword{datasets}
