% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_2.R
\docType{data}
\name{table_6_2}
\alias{table_6_2}
\title{Table 6.2 Gross Private Domestic Investment and GDP, United States, 1990-2005}
\format{
A \link[tibble:tibble-package]{tibble} with 16 rows and
5 variables:
\itemize{
\item \strong{year}
\item \strong{gpdibl:} gross private domestic investment, billions of 2000 dollars
\item \strong{gpdim:} gross private domestic investment, millions of 2000 dollars
\item \strong{gdpb:} gross domestic product, billions of 2000 dollars
\item \strong{gdpm:} gross domestic product, millions of 2000 dollars
}
}
\source{
Economic Report of the President, 2007, Table B-2, p. 328.
}
\usage{
table_6_2
}
\description{
Gross Private Domestic Investment and GDP, United States, 1990–2005
(Billions of chained - 2000 - dollars, except as noted; quarterly data at
seasonally adjusted annual rates)
}
\keyword{datasets}
