% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_3.R
\docType{data}
\name{table_6_3}
\alias{table_6_3}
\title{Table 6.3 Total Personal Expenditure and Categories}
\format{
A \link[tibble:tibble-package]{tibble} with 15 rows and
5 variables:
\describe{
\item{year_or_quarter}{}
\item{expservices:}{expenditure on services, billions of 2000 dollars}
\item{expdur:}{expenditure on durable goods, billions of 2000 dollars}
\item{expnondur:}{expenditure on nondurable goods, billions of 2000 dollars}
\item{pcexp:}{total personal consumption expenditure, billions of 2000 dollars}
}
}
\source{
Department of Commerce, Bureau of Economic Analysis. Economic Report
of the President, 2007, Table B-17, p. 347.
}
\usage{
table_6_3
}
\description{
Total Personal Expenditure and Categories
(Billions of chained - 2000 - dollars; quarterly data at seasonally adjusted
annual rates)
}
\keyword{datasets}
