% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_4.R
\docType{data}
\name{table_6_4}
\alias{table_6_4}
\title{Table 6.4 Fertility and Other Data for 64 Countries}
\format{
A \link[tibble:tibble-package]{tibble} with 64 rows and
5 variables:
\itemize{
\item \strong{obs}
\item \strong{cm:} child mortality, the number of deaths of children under age 5 in year per 1000 live births
\item \strong{flfp:} female literacy rate, percent
\item \strong{pgnp:} per capita GNP in 1980
\item \strong{tfr:} total fertility rate, 1980-1985, the average number of children born to a woman, using age-specific fertility rates for a given year
}
}
\source{
Chandan Mukherjee, Howard White, and Marc Whyte, Econometrics and
Data Analysis for Developing Countries, Routledge, London, 1998, p. 456.
}
\usage{
table_6_4
}
\description{
Fertility and Other Data for 64 Countries
}
\keyword{datasets}
