% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_5.R
\docType{data}
\name{table_6_5}
\alias{table_6_5}
\title{Table 6.5 Inﬂation Rate and Unemployment Rate, United States, 1960–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 47 rows and
3 variables:
\describe{
\item{year}{}
\item{inflrate:}{inflation rate, percent year-to-year chance in CPI}
\item{unrate:}{unemployment rate, the civilian unemployment rate}
}
}
\source{
Economic Report of the President, 2007, Table B-60, p. 399,
for CPI changes and Table B-42, p. 376, for the unemployment rate
}
\usage{
table_6_5
}
\description{
Table 6.5 Inﬂation Rate and Unemployment Rate, United States, 1960–2006
(For all urban consumers; 1982–1984 = 100, except as noted)
}
\keyword{datasets}
