% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_7.R
\docType{data}
\name{table_6_7}
\alias{table_6_7}
\title{Table 6.7}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
2 variables:
\describe{
\item{y}{}
\item{x}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 178. Adapted from J. Johnston, Econometric
Methods, 3d ed., McGraw-Hill, New York, 1984, p. 87. Actually this is taken
from an econometric examination of Oxford University in 1975.
}
\usage{
table_6_7
}
\description{
Table 6.7
}
\keyword{datasets}
