% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_8.R
\docType{data}
\name{table_6_8}
\alias{table_6_8}
\title{Table 6.8 Relationship between investment rate and savings rate}
\format{
A \link[tibble:tibble-package]{tibble} with 21 rows and
3 variables:
\itemize{
\item \strong{country}
\item \strong{savrate:} savings as a ratio of GDP
\item \strong{invrate:} investment expenditure as a ration of GDP
}
}
\source{
FELDSTEIN, Martin; HORIOKA, Charles. “Domestic Saving and International Capital Flows,”
Economic Journal, vol. 90, June 1980, pp. 314–329. Data reproduced from Michael P. Murray,
Econometrics: A Modern Introduction, Addison-Wesley, Boston, 2006.
}
\usage{
table_6_8
}
\description{
Relationship between investment rate (investment expenditure as a ratio of the GDP)
and savings rate (savings as a ratio of GDP)
}
\keyword{datasets}
