% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_1.R
\docType{data}
\name{table_7_1}
\alias{table_7_1}
\title{Table 7.1 U.S. Coffee Consumption in Relation to Average Real Retail Price}
\format{
A \link[tibble:tibble-package]{tibble} with 11 rows and
3 variables:
\describe{
\item{year}{}
\item{y:}{cups per person per day}
\item{x:}{average real retail price, $ per pound}
}
}
\source{
The data for Y are from Summary of National Coffee Drinking Study,
Data Group, Elkins Park, Penn., 1981; and the data on nominal X
(i.e., X in current prices) are from Nielsen Food Index, A. C. Nielsen, New York, 1981.
}
\usage{
table_7_1
}
\description{
U.S. Coffee Consumption (Y) in Relation to Average Real Retail Price (X), 1970–1980.
Note: The nominal price was divided by the Consumer Price Index (CPI) for food and beverages, 1967 = 100.
}
\keyword{datasets}
