% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_10.R
\docType{data}
\name{table_7_10}
\alias{table_7_10}
\title{Table 7.10 Demand for Money in the United States, 1980–1998}
\format{
A \link[tibble:tibble-package]{tibble} with 19 rows and
6 variables:
\describe{
\item{obs:}{observation, year}
\item{gdp:}{gross domestic product ($ billions)}
\item{m2:}{M2 money supply}
\item{cpi:}{costumer price index (1982-1984=100)}
\item{ltrate:}{long-term interest rate (30-year treasury bond)}
\item{tbrate:}{three-month treasury bill rate (\% per annum)}
}
}
\source{
Economic Report of the President, 2000, Tables B-1, B-58, B-67, B-71.
}
\usage{
table_7_10
}
\description{
Demand for Money in the United States, 1980–1998
}
\keyword{datasets}
