% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_11.R
\docType{data}
\name{table_7_11}
\alias{table_7_11}
\title{Table 7.11 Greek Industrial Sector}
\format{
A \link[tibble:tibble-package]{tibble} with 27 rows and
5 variables:
\itemize{
\item \strong{year}
\item \strong{output:} billions of Drachmas at constant 1970 prices
\item \strong{capital}
\item \strong{labor:} thousands of workers per year
\item \strong{ratio:} capital-to-labor
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 224. I am indebted to George K. Zestos
of Christopher Newport University, Virginia, for these data.
}
\usage{
table_7_11
}
\description{
Table 7.11 Greek Industrial Sector
}
\keyword{datasets}
