% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_12.R
\docType{data}
\name{table_7_12}
\alias{table_7_12}
\title{Table 7.17 Real Consumption Expenditure, Real Income, Real Wealth, and Real Interest Rates for the U.S., 1947–2000}
\format{
A \link[tibble:tibble-package]{tibble} with 54 rows and
5 variables:
\describe{
\item{year:}{calendar year}
\item{c:}{real consumption expenditure in billions of chained 1996 dollars}
\item{yd:}{real personal disposable income in billions of chained 1996 dollars}
\item{wealth:}{real wealth in billions of chained 1996 dollars}
\item{interest_rate:}{nominal annual yield on 3-month Treasury securities–inﬂation rate (measured by the annual \% change in annual chained price index)}
}
}
\source{
C, Yd, and quarterly and annual chain-type price indexes (1996 = 100):
Bureau of Economic Analysis, U.S. Department of Commerce \url{http://www.bea.doc.gov/bea/dn1.htm}
Nominal annual yield on 3-month Treasury securities: Economic Report of the President, 2002.
Nominal wealth = end-of-year nominal net worth of households and nonproﬁts
(from Federal Reserve ﬂow of funds data: \url{http://www.federalreserve.gov}
}
\usage{
table_7_12
}
\description{
Real Consumption Expenditure, Real Income, Real Wealth, and Real Interest Rates for the U.S., 1947–2000
}
\keyword{datasets}
