% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_3.R
\docType{data}
\name{table_7_3}
\alias{table_7_3}
\title{Table 7.3 Value Added, Labor Hours, and Capital Input in the Manufacturing Sector of the U.S., 2005}
\format{
A \link[tibble:tibble-package]{tibble} with 51 rows and
4 variables:
\describe{
\item{area}{}
\item{y:}{output value added (thousands of $)}
\item{x1:}{labor input worker hrs (thousands)}
\item{x2:}{capital input capital expenditure (thousands of $)}
}
}
\source{
2005 Annual Survey of Manufacturers, Sector 31: Supplemental Statistics for U.S.
}
\usage{
table_7_3
}
\description{
Value Added, Labor Hours, and Capital Input in the Manufacturing Sector of the U.S., 2005
}
\keyword{datasets}
