% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_4.R
\docType{data}
\name{table_7_4}
\alias{table_7_4}
\title{Table 7.4 Total Cost (Y) and Output (X)}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
2 variables:
\itemize{
\item \strong{x:} output
\item \strong{y:} total cost, $
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 211.
}
\usage{
table_7_4
}
\description{
Total Cost (Y) and Output (X)
}
\keyword{datasets}
