% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_5.R
\docType{data}
\name{table_7_5}
\alias{table_7_5}
\title{Table 7.5}
\format{
A \link[tibble:tibble-package]{tibble} with 3 rows and
3 variables:
\describe{
\item{y}{}
\item{x2}{}
\item{x3}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 216.
}
\usage{
table_7_5
}
\description{
Table 7.5
}
\keyword{datasets}
