% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_6.R
\docType{data}
\name{table_7_6}
\alias{table_7_6}
\title{Table 7.6 Quarterly Demand for Roses in Metro Detroit Area, from 1971-III to 1975-II}
\format{
A \link[tibble:tibble-package]{tibble} with 16 rows and
6 variables:
\itemize{
\item \strong{year_quarter}
\item \strong{y:} quantity of roses sold, dozens
\item \strong{x2:} average wholesale price of roses, $/dozen
\item \strong{x3:} average wholesale price of carnations, $/dozen
\item \strong{x4:} average weekly family disposable income, $/week
\item \strong{x5:} the trend variable taking values of 1, 2, and so on, for the perior 1971-III to 1975-II in the Detroit metropolitan area
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 219
}
\usage{
table_7_6
}
\description{
Quarterly Demand for Roses in Metro Detroit Area, from 1971-III to 1975-II
}
\keyword{datasets}
