% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8_10.R
\docType{data}
\name{table_8_10}
\alias{table_8_10}
\title{Table 8.10 The demand for cable}
\format{
A \link[tibble:tibble-package]{tibble} with 16 rows and
7 variables:
\describe{
\item{year}{}
\item{x2:}{gross national product (GNP), $ billions}
\item{x3:}{housing starts, thousands of units}
\item{x4:}{unemployment rate, \%}
\item{x5:}{prime rate lag, 6 months}
\item{x6:}{customer line gains, \%}
\item{y:}{annual sales in MPF, million paired feet}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 269.
I am indebted to Daniel J. Reardon for collecting and processing the data.
}
\usage{
table_8_10
}
\description{
Data used by a telephone cable manufacturer to predict sales to a major customer for the period 1968–1983
}
\keyword{datasets}
