% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8_11.R
\docType{data}
\name{table_8_11}
\alias{table_8_11}
\title{Table 8.11 Savings and Personal Disposable Income, United States, 1970–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
3 variables:
\itemize{
\item \strong{year}
\item \strong{savings}
\item \strong{income}
}
}
\source{
Department of Commerce, Bureau of Economic Analysis.
}
\usage{
table_8_11
}
\description{
Savings and Personal Disposable Income (billions of dollars), United States,
1970–2005 (billions of dollars, except as noted; quarterly data at seasonally adjusted annual rates)
}
\keyword{datasets}
