% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8_8.R
\docType{data}
\name{table_8_8}
\alias{table_8_8}
\title{Table 8.8 Real GDP, Employment, and Real Fixed Capital—Mexico}
\format{
A \link[tibble:tibble-package]{tibble} with 20 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{gdp:} millions of 1960 pesos
\item \strong{employment:} thousands of people
\item \strong{capital:} fixed capital, millions of 1960 pesos
}
}
\source{
Victor J. Elias, Sources of Growth: A Study of Seven Latin American Economies,
International Center for Economic Growth, ICS Press, San Francisco, 1992.
Data from Tables E5, E12, and E14.
}
\usage{
table_8_8
}
\description{
Real GDP, Employment, and Real Fixed Capital—Mexico
}
\keyword{datasets}
