% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_8_9.R
\docType{data}
\name{table_8_9}
\alias{table_8_9}
\title{Table 8.9 Savings and Personal Disposable Income (billions of dollars), United States, 1970–1995}
\format{
A \link[tibble:tibble-package]{tibble} with 26 rows and
3 variables:
\itemize{
\item \strong{year}
\item \strong{savings}
\item \strong{income}
}
}
\source{
Economic Report of the President, 1997, Table B-28, p. 332.
}
\usage{
table_8_9
}
\description{
Savings and Personal Disposable Income (billions of dollars), United States, 1970–1995
}
\keyword{datasets}
