% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_1.R
\docType{data}
\name{table_9_1}
\alias{table_9_1}
\title{Table 9.1 Average Salary of Public School Teachers by State, 2005–2006}
\format{
A \link[tibble:tibble-package]{tibble} with 51 rows and
5 variables:
\itemize{
\item \strong{state}
\item \strong{salary}
\item \strong{spending}
\item \strong{d2:} 1 for states in the Northeast and North Central; 0 otherwise
\item \strong{d3:} 1 for states in the South; 0 otherwise
}
}
\source{
National Educational Association, as reported in 2007.
}
\usage{
table_9_1
}
\description{
Average Salary of Public School Teachers by State, 2005–2006
}
\keyword{datasets}
