% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_2.R
\docType{data}
\name{table_9_2}
\alias{table_9_2}
\title{Table 9.2 Savings and Income Data, United States, 1970–1995}
\format{
A \link[tibble:tibble-package]{tibble} with 26 rows and
4 variables:
\describe{
\item{year}{}
\item{savings:}{billions of dollars}
\item{income:}{billions of dollars}
\item{dum:}{1 for observations beginning in 1982; 0 otherwise}
}
}
\source{
Economic Report of the President, 1997, Table B-28, p. 332.
}
\usage{
table_9_2
}
\description{
Savings and Income Data, United States, 1970–1995
}
\keyword{datasets}
