% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_3.R
\docType{data}
\name{table_9_3}
\alias{table_9_3}
\title{Table 9.3 Quarterly Data on Appliance Sales and Expenditure on Durable Goods}
\format{
A \link[tibble:tibble-package]{tibble} with 32 rows and
5 variables:
\describe{
\item{dish:}{dishwashers}
\item{disp:}{garbage disposers}
\item{frig:}{refrigerators}
\item{wash:}{washing machines}
\item{dur:}{durable goods expenditure, billions of 1982 dollas}
}
}
\source{
Business Statistics and Survey of Current Business, Department of Commerce (various issues).
}
\usage{
table_9_3
}
\description{
Quarterly Data on Appliance Sales (in thousands) and Expenditure on Durable Goods (1978–I to 1985–IV)
}
\keyword{datasets}
