% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_4.R
\docType{data}
\name{table_9_4}
\alias{table_9_4}
\title{Table 9.4 U.S. Refrigerator Sales, 1978–1985}
\format{
A \link[tibble:tibble-package]{tibble} with 32 rows and
5 variables:
\describe{
\item{frig:}{refrigerator sales, thousands}
\item{dur:}{durable goods expenditure, billions of 1982 dollars}
\item{d2:}{1 in the second quarter, 0 otherwise}
\item{d3:}{1 in the third quarter, 0 otherwise}
\item{d4:}{1 in the fourth quarter, 0 otherwise}
}
}
\source{
Business Statistics and Survey of Current Business, Department of Commerce (various issues).
}
\usage{
table_9_4
}
\description{
U.S. Refrigerator Sales (thousands), 1978–1985 (quarterly)
}
\keyword{datasets}
