% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_6.R
\docType{data}
\name{table_9_6}
\alias{table_9_6}
\title{Table 9.6 Hypothetical Data on Output and Total Cost}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
2 variables:
\itemize{
\item \strong{total_cost:} dollars
\item \strong{output:} units
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 297.
}
\usage{
table_9_6
}
\description{
Hypothetical Data on Output and Total Cost
}
\keyword{datasets}
