% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_7.R
\docType{data}
\name{table_9_7}
\alias{table_9_7}
\title{Table 9.7 Indian Wage Earners, 1990}
\format{
A \link[tibble:tibble-package]{tibble} with 114 rows and
7 variables:
\itemize{
\item \strong{wi:} weekly wage income in rupees
\item \strong{age:} age in years
\item \strong{de2:} a dummy variable taking a value of 1 for workers with an education level up to primary
\item \strong{de3:} a dummy variable taking a value of 1 for workers up to a secondary level of education
\item \strong{de4:} a dummy variable taking a value of 1 for workers with higher than secondary education
\item \strong{dpt:} a dummy variable taking a value of 1 for workers with permanent jobs and a value of 0 for temporary workers
\item \strong{dsex:} 1 for male workers and 0 for female workers
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 300-301.
}
\usage{
table_9_7
}
\description{
Data on a sample of 261 workers in an industrial town in southern India in 1990.
}
\keyword{datasets}
