% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_9_8.R
\docType{data}
\name{table_9_8}
\alias{table_9_8}
\title{Table 9.8 Data Matrix for Regression}
\format{
A \link[tibble:tibble-package]{tibble} with 51 rows and
5 variables:
\describe{
\item{year_quarter:}{year and quarter}
\item{un:}{unemployment rate, \%}
\item{v:}{job vacancy rate, \%}
\item{d:}{1, for period beginning in 1966-IV}
\item{dv:}{v when d = 1}
}
}
\source{
Damodar Gujarati, “The Behaviour of Unemployment and Unﬁlled Vacancies:
Great Britain, 1958–1971,” The Economic Journal, vol. 82, March 1972, p. 202.
}
\usage{
table_9_8
}
\description{
Data Matrix for Regression
}
\keyword{datasets}
