% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_e_1.R
\docType{data}
\name{table_e_1}
\alias{table_e_1}
\title{Table E.1 U.S. Civilian Labor Force Participation Rate, Civilian Unemployment Rate and Real Average Hourly Earnings for the years 1980-2002}
\format{
A \link[tibble:tibble-package]{tibble} with 23 rows and
4 variables:
\describe{
\item{year}{}
\item{clfpr:}{civilian labor force participation rate}
\item{cnur:}{civilian unemployment rate}
\item{ahe82:}{real average hourly earnings}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 894.
}
\usage{
table_e_1
}
\description{
U.S. Civilian Labor Force Participation Rate, Civilian Unemployment Rate and Real Average Hourly Earnings for the years 1980-2002
}
\keyword{datasets}
