% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_i_1.R
\docType{data}
\name{table_i_1}
\alias{table_i_1}
\title{Table I.1 Personal Consumption Expenditure x Gross Domesctic Product}
\format{
A \link[tibble:tibble-package]{tibble} with 46 rows and
3 variables:
\itemize{
\item \strong{year}
\item \strong{pce_y:} personal consumption expenditure
\item \strong{gdp_x:} gross domestic product
}
}
\source{
Economic Report of the President, 2007, Table B-2, p. 230.
}
\usage{
table_i_1
}
\description{
Data on Y (Personal Consumption Expenditure) and X (Gross Domesctic Product, 1960-2005),
both in 2000 Billions of Dollars
}
\keyword{datasets}
