% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableE_1.R
\docType{data}
\name{TableE_1}
\alias{TableE_1}
\title{Table E_1}
\format{

}
\usage{
data('TableE_1')
}
\description{
Source: See the page 894 to more information
}
\note{
U.S. economy for the period 1980 to 2002
\itemize{
\item \strong{Year}
\item \strong{CLFPR: }civilian labor force participation rate
\item \strong{CUNR: }civilian unemployment rate
\item \strong{AHE82: }real average hourly earnings in 1982 dollars (AHE82)
}
}
\keyword{datasets}
