% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableI_1.R
\docType{data}
\name{TableI_1}
\alias{TableI_1}
\title{Table I_1}
\format{


A data.frame with 46 observations on 3 variables
\itemize{
\item \strong{Year}
\item \strong{Y:} Personal Consumption Expenditure in Billions of Dollars
\item \strong{X:} Gross Domestic Product in Billions of Dollars
}
}
\usage{
data('TableI_1')
}
\description{
DATA ON Y (PERSONAL CONSUMPTION EXPENDITURE) AND X (GROSS DOMESTIC PRODUCT, 1982 1996), BOTH IN 1992 BILLIONS OF DOLLARS
Source: DATA ON Y (PERSONAL CONSUMPTION EXPENDITURE) AND X (GROSS DOMESTIC PRODUCT, 1982 1996), BOTH IN 1992 BILLIONS OF DOLLARS
}
\keyword{datasets}
