% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_11.R
\docType{data}
\name{table_10_11}
\alias{table_10_11}
\title{Table 10.11 Hypothetical Data}
\format{
A \link[tibble:tibble-package]{tibble} with 11 rows and
3 variables:
\describe{
\item{y}{}
\item{x2}{}
\item{x3}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 352.
}
\usage{
table_10_11
}
\description{
Hypothetical Data
}
\keyword{datasets}
