% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_13.R
\docType{data}
\name{table_10_13}
\alias{table_10_13}
\title{Table 10.13 U.S. Imports, GDP, and CPI, 1975–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 31 rows and
4 variables:
\itemize{
\item \strong{year}
\item \strong{cpi:} consumer price index
\item \strong{gdp:} gross domestic product
\item \strong{imports}
}
}
\source{
Department of Labor, Bureau of Labor Statistics.
}
\usage{
table_10_13
}
\description{
U.S. Imports, GDP, and CPI, 1975–2005
(For all urban consumers; 1982–84 = 100, except as noted)
}
\keyword{datasets}
