% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_10_15.R
\docType{data}
\name{table_10_15}
\alias{table_10_15}
\title{Table 10.15 Hours of Work and Other Data for 35 Groups}
\format{
A \link[tibble:tibble-package]{tibble} with 35 rows and
10 variables:
\describe{
\item{obs:}{observation}
\item{hours:}{average hours worked during the year}
\item{rate:}{average hourly wage (dollars)}
\item{ersp:}{ average yearly earnings of spouse (dollars)}
\item{erno:}{average yearly earnings of other family members (dollars)}
\item{nein:}{average yearly nonearned income}
\item{assets:}{average family asset holdings (bank account, etc.) (dollars)}
\item{age:}{average age of respondent}
\item{dep:}{average number of dependents}
\item{school:}{average highest grade of school completed}
}
}
\source{
D. H. Greenberg and M. Kosters, Income Guarantees and the Working Poor,
Rand Corporation, R-579-OEO, December 1970
}
\usage{
table_10_15
}
\description{
Hours of Work and Other Data for 35 Groups
}
\keyword{datasets}
