% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_1.R
\docType{data}
\name{table_11_1}
\alias{table_11_1}
\title{Table 11.1 Compensation per Employee ($) in Nondurable Manufacturing Industries According to Employment Size of Establishment, 1958}
\format{
A \link[tibble:tibble-package]{tibble} with 13 rows and
11 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{industry}
\item \strong{a:} employment size (average number of employees) = 1-4
\item \strong{b:} employment size (average number of employees) = 5-9
\item \strong{c:} employment size (average number of employees) = 10-19
\item \strong{d:} employment size (average number of employees) = 20-49
\item \strong{e:} employment size (average number of employees) = 50-99
\item \strong{f:} employment size (average number of employees) = 100-249
\item \strong{g:} employment size (average number of employees) = 250-499
\item \strong{h:} employment size (average number of employees) = 500-999
\item \strong{i:} employment size (average number of employees) = 1,000-2,499
}
}
\source{
The Census of Manufacturers, U.S. Department of Commerce, 1958 (computed by author).
}
\usage{
table_11_1
}
\description{
Compensation per Employee ($) in Nondurable Manufacturing Industries According to Employment Size of Establishment, 1958
}
\keyword{datasets}
