% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_10.R
\docType{data}
\name{table_11_10}
\alias{table_11_10}
\title{Table 11.10 Salary and related data on 447 executives of fortune 500 companies}
\format{
A \link[tibble:tibble-package]{tibble} with 447 rows and
7 variables:
\describe{
\item{salary:}{1999 salary and bonuses}
\item{totcomp:}{1999 CEO total compensation}
\item{tenure:}{number of years as CEO (0 if less than 6 months)}
\item{age:}{age of CEO}
\item{sales:}{total 1998 sales revenue of the firm}
\item{profits:}{1998 profits for the firm}
\item{assets:}{total assets of the firm in 1998}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 409.
}
\usage{
table_11_10
}
\description{
Salary and related data on 447 executives of fortune 500 companies.
}
\keyword{datasets}
