% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_11_6.R
\docType{data}
\name{table_11_6}
\alias{table_11_6}
\title{Table 11.6 Asset Size}
\format{
A \link[tibble:tibble-package]{tibble} with 12 rows and
8 variables:
\describe{
\item{year_quarter:}{}
\item{a:}{1-10}
\item{b:}{10-25}
\item{c:}{25-50}
\item{d:}{50-100}
\item{e:}{100-250}
\item{f:}{250-1,000}
\item{g:}{1,000 +}
}
}
\source{
Quarterly Financial Report for Manufacturing Corporations,
Federal Trade Commission and the Securities and Exchange Commission,
U.S. government, various issues (computed).
}
\usage{
table_11_6
}
\description{
Asset Size (millions of dollars)
}
\keyword{datasets}
