% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_12_10.R
\docType{data}
\name{table_12_10}
\alias{table_12_10}
\title{Table 12.10 Rate of Return, Output Growth and Inﬂation, United States, 1954–1981}
\format{
A \link[tibble:tibble-package]{tibble} with 28 rows and
4 variables:
\describe{
\item{obs:}{observation; year}
\item{rr:}{real rate of return on common stocks at time, \%}
\item{growth:}{output growth in period (t + 1), \%}
\item{inflation:}{inflation in period t(Inft), \%}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 464.
}
\usage{
table_12_10
}
\description{
Rate of Return, Output Growth and Inﬂation, United States, 1954–1981
}
\keyword{datasets}
