% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_12_7.R
\docType{data}
\name{table_12_7}
\alias{table_12_7}
\title{Table 12.7 Determinants of U.S. Domestic Price of Copper, 1951–1980}
\format{
A \link[tibble:tibble-package]{tibble} with 30 rows and
7 variables:
\describe{
\item{year}{}
\item{c:}{12-month average U.S. domestic price of copper (cents per pound)}
\item{g:}{annual gross national product ($, billions)}
\item{i:}{12-month average index of industrial production}
\item{l:}{12-month average London Metal Exchange price of copper (pounds sterling)}
\item{h:}{number of housing starts per year (thousands of units)}
\item{a:}{12-month average price of aluminum (cents per pound)}
}
}
\source{
The data were collected by Gary R. Smith from sources such as
American Metal Market, Metals Week, and U.S. Department of Commerce publications.
}
\usage{
table_12_7
}
\description{
Determinants of U.S. Domestic Price of Copper, 1951–1980
}
\keyword{datasets}
