% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_13_2.R
\docType{data}
\name{table_13_2}
\alias{table_13_2}
\title{Table 13.2 Hypothetical Data on True Consumption Expenditure, True Income, Measured Consumption Expenditure and Measured Income}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
7 variables:
\describe{
\item{y_star:}{true Consumption Expenditure, $}
\item{x_star:}{true Income, $}
\item{y:}{measure Consumption Expenditure, $}
\item{x:}{measured Income, $}
\item{e:}{stochastic Error Term}
\item{w:}{stochastic Error Term}
\item{u:}{stochastic Error Term}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 485.
}
\usage{
table_13_2
}
\description{
Hypothetical Data on True Consumption Expenditure, True Income, Measured Consumption Expenditure and Measured Income
}
\keyword{datasets}
