% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_23.R
\docType{data}
\name{table_15_23}
\alias{table_15_23}
\title{Table 15.23 GRE Scores}
\format{
A \link[tibble:tibble-package]{tibble} with 13 rows and
4 variables:
\itemize{
\item \strong{student:} student number
\item \strong{q:} GRE aptitude test score - quantitative
\item \strong{v:} GRE aptitude test score - verbal
\item \strong{adm:} admitted to graduate program; = 1 if yes, = 0 if no
}
}
\source{
Donald F. Morrison, Applied Linear Statistical Methods, Prentice-Hall, Inc.,
Englewood Cliffs, NJ, 1983, p. 279 (adapted).
}
\usage{
table_15_23
}
\description{
Applicants to a Graduate Program
}
\keyword{datasets}
