% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_27.R
\docType{data}
\name{table_15_27}
\alias{table_15_27}
\title{Table 15.27 Data for 2,000 women regarding work}
\format{
A \link[tibble:tibble-package]{tibble} with 2000 rows and
15 variables:
\itemize{
\item \strong{c1}
\item \strong{c2}
\item \strong{u}
\item \strong{v}
\item \strong{county}
\item \strong{age}
\item \strong{education:} number of years of schooling
\item \strong{married:} = 1, married; = 0, otherwise
\item \strong{children:} number of children
\item \strong{select}
\item \strong{wagefull}
\item \strong{wage}
\item \strong{lw}
\item \strong{work} = 1, woman works; = 0, otherwise
\item \strong{lwf}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 589.
}
\usage{
table_15_27
}
\description{
Data for 2,000 women regarding work
}
\keyword{datasets}
