% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_29.R
\docType{data}
\name{table_15_29}
\alias{table_15_29}
\title{Table 15.29}
\format{
A \link[tibble:tibble-package]{tibble} with 178 rows and
12 variables:
\describe{
\item{cancer:}{cancer = 1, had breast cancer; cancer = 0, did not have breast cancer}
\item{age:}{age of subject}
\item{higd:}{highest grade completed in school}
\item{degree}{}
\item{chk:}{chk = 0 if subject did not undergo regular medical checkups; chk = 1 if subject did undergo regular checkups}
\item{agpi:}{age at first pregnancy}
\item{men_age}{}
\item{miscarriages:}{number of miscarriages}
\item{births}{}
\item{weight:}{weight of subject}
\item{age_period}{}
\item{marital_status}{}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 589.
}
\usage{
table_15_29
}
\description{
Predicting whether a woman will contract breast cancer or not.
}
\keyword{datasets}
