% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_15_7.R
\docType{data}
\name{table_15_7}
\alias{table_15_7}
\title{Table 15.7 Effect of Personalized System of Instruction (PSI) on Course Grades}
\format{
A \link[tibble:tibble-package]{tibble} with 32 rows and
6 variables:
\itemize{
\item \strong{obs:} observation
\item \strong{gpa:} entering grade point average
\item \strong{tuce:} score on an examination given at the beginning of the term to test entering knowledge of macroeconomics
\item \strong{psi:} personal system of instruction; = 1 if the new method is used, = 0 otherwise
\item \strong{grade:} = 1 if the letter grade is A, = 0 otherwise
\item \strong{letter:} letter grade
}
}
\source{
L. Spector and M. Mazzeo, “Probit Analysis and Economic Education,”
Journal of Economic Education, vol. 11, 1980, pp. 37–44.
}
\usage{
table_15_7
}
\description{
Data on the Effect of Personalized System of Instruction (PSI) on Course Grades
}
\keyword{datasets}
