% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_16_18.R
\docType{data}
\name{table_16_18}
\alias{table_16_18}
\title{Table 16.18 Responsiveness of aid to recipient countries’ economic and physical needs, civil/political rights, and government effectiveness}
\format{
A \link[tibble:tibble-package]{tibble} with 395 rows and
24 variables:
\itemize{
\item \strong{countryid}
\item \strong{country}
\item \strong{cntry_code}
\item \strong{year}
\item \strong{aidcap}
\item \strong{raidcap}
\item \strong{infmort}
\item \strong{infmort_2}
\item \strong{pop}
\item \strong{raid}
\item \strong{civlib}
\item \strong{polright}
\item \strong{civil_polrts}
\item \strong{goveff}
\item \strong{regqual}
\item \strong{rgdpcap1}
\item \strong{rgdpcap}
\item \strong{rgdcap_2}
\item \strong{popmil}
\item \strong{popmil_2}
\item \strong{deflator}
\item \strong{y1}
\item \strong{y2}
\item \strong{y3}
}
}
\source{
Subhayu Bandyopadhyay and Howard J. Wall, “The Determinants of Aid in
the Post-Cold War Era,” Review,  Federal Reserve Bank of St. Louis,
November/December 2007, vol. 89, number 6, pp. 533–547.
The article and data can be found at: \url{http://research.stlouisfed.org/publications/review/past/2007}
in the November/December, Vol. 89, No. 10 section.
}
\usage{
table_16_18
}
\description{
Data to estimate the responsiveness of aid to recipient countries’ economic and physical needs,
civil/political rights, and government effectiveness. The data are for 135 countries for three years.
}
\keyword{datasets}
