% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_13.R
\docType{data}
\name{table_17_13}
\alias{table_17_13}
\title{Table 17.13 Macroeconomic Data for the Greek Economy, 1960–1995}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
6 variables:
\itemize{
\item \strong{year}
\item \strong{pc:} real desired private consumption expenditure
\item \strong{pdi:} real private disposable income
\item \strong{grossinv}
\item \strong{gnp}
\item \strong{lti:} long-term interest rate
}
}
\source{
H. R. Seddighi, K. A. Lawler, and A. V. Katos, Econometrics:
A Practical Approach, Routledge, London, 2000, p. 158.
}
\usage{
table_17_13
}
\description{
Macroeconomic Data for the Greek Economy, 1960–1995
}
\keyword{datasets}
