% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_17_5.R
\docType{data}
\name{table_17_5}
\alias{table_17_5}
\title{Table 17.5 Money, Interest Rate, Price Index, and GDP, Canada}
\format{
A \link[tibble:tibble-package]{tibble} with 40 rows and
5 variables:
\describe{
\item{obs:}{observation; year and quarter}
\item{m1:}{C$, millions}
\item{r:}{90-day prime corporate interest rate, \%}
\item{p:}{implicit price deflator (1981 = 100)}
\item{gdp:}{C$, millions (1981 prices)}
}
}
\source{
Rao, op. cit., pp. 210–213.
}
\usage{
table_17_5
}
\description{
Money, Interest Rate, Price Index, and GDP, Canada
}
\keyword{datasets}
