% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_18_4.R
\docType{data}
\name{table_18_4}
\alias{table_18_4}
\title{Table 18.4 Macroeconomic data on several variables for the U.S. economy for the quarterly periods 1951–I to 2000–IV}
\format{
A \link[tibble:tibble-package]{tibble} with 204 rows and
14 variables:
\describe{
\item{year:}{date}
\item{qtr:}{quarter}
\item{realgdp:}{real GDP (billions of dollars)}
\item{realcons:}{real consumption expenditure}
\item{realinvs:}{real investment by private sector}
\item{realgovt:}{real government expenditure}
\item{realdpi:}{real disposable personal income}
\item{cpi_u:}{consumer price index}
\item{m1:}{nominal money stock}
\item{tbilrate:}{quarterly average of month-end 90-day T-bill rate}
\item{unemp}{}
\item{pop:}{population, millions, interpolate of year-end ﬁgures using constant growth rate per quarter}
\item{infl:}{rate of inﬂation (ﬁrst observation is missing}
\item{realint:}{expost real interest rate}
}
}
\source{
Department of Commerce, Bureau of Economic Analysis, and from \url{www.economagic.com},
and are reproduced from William H. Greene, Econometric Analysis, 6th ed., 2008, Table F5.1, p.1083.
}
\usage{
table_18_4
}
\description{
Macroeconomic data on several variables for the U.S. economy for the quarterly periods 1951–I to 2000–IV
}
\keyword{datasets}
