% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_20_2.R
\docType{data}
\name{table_20_2}
\alias{table_20_2}
\title{Table 20.2 GDP, M2, FEDEXP, TB6, USA, 1970–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 36 rows and
6 variables:
\describe{
\item{obs:}{observation; year}
\item{y1:}{GDP = gross domestic product (billions of chained 2000 dollars)}
\item{y2:}{M2 = M2 money supply (billions of dollars)}
\item{x1:}{GPDI = gross private domestic investment (billions of chained 2000 dollars)}
\item{x2:}{FEDEXP = Federal government expenditure (billions of dollars)}
\item{x3:}{TB6 = 6-month Treasury bill rate (\%)}
}
}
\source{
Economic Report of the President, 2007. Tables B-2, B-69, B-84, and B-73.
}
\usage{
table_20_2
}
\description{
GDP, M2, FEDEXP, TB6, USA, 1970–2005
}
\keyword{datasets}
