% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_20_5.R
\docType{data}
\name{table_20_5}
\alias{table_20_5}
\title{Table 20.5 Underlying Data for Klein’s Model I}
\format{
A \link[tibble:tibble-package]{tibble} with 22 rows and
10 variables:
\itemize{
\item \strong{year}
\item \strong{c:} consumption expenditure
\item \strong{p:} profits
\item \strong{w:} private wage bill
\item \strong{i:} investment expenditure
\item \strong{k-1:} capital stock
\item \strong{x:} income after tax
\item \strong{w':} government wage bill
\item \strong{g:} government expenditure
\item \strong{t:} taxes
}
}
\source{
G. S. Maddala, Econometrics, McGraw-Hill, New York, 1977, p. 238.
}
\usage{
table_20_5
}
\description{
Underlying Data for Klein’s Model I
}
\keyword{datasets}
