% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_10.R
\docType{data}
\name{table_2_10}
\alias{table_2_10}
\title{Table 2.10 SAT Reasoning Test Classified by Family Income}
\format{
A \link[tibble:tibble-package]{tibble} with 10 rows and
9 variables:
\itemize{
\item \strong{family_income}
\item \strong{middle_income:} middle income in the interval
\item \strong{n_test_takers}
\item \strong{cr_mean:} critical reading mean
\item \strong{cr_sd:} critical reading standart deviation
\item \strong{m_mean:} mathematics mean
\item \strong{m_sd:} mathematics standart deviation
\item \strong{w_mean:} writing mean
\item \strong{w_sd:} writing standart deviation
}
}
\source{
College Board, 2007. College-Bound Seniors, Table 11.
}
\usage{
table_2_10
}
\description{
SAT Reasoning Test Classified by Family Income
}
\keyword{datasets}
