% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_2_6.R
\docType{data}
\name{table_2_6}
\alias{table_2_6}
\title{Table 2.6 Mean Hourly Wage by Education}
\format{
A \link[tibble:tibble-package]{tibble} with 13 rows and
3 variables:
\describe{
\item{years_schooling:}{years of schooling}
\item{mean_wage:}{mean wage, $}
\item{n_people:}{number of people}
}
}
\source{
Arthur S. Goldberger. Introductory Econometrics, Harvard University Press,
Cambridge, Mass., 1998, Table 1.1, p.5 (adapted).
}
\usage{
table_2_6
}
\description{
Mean Hourly Wage by Education
}
\keyword{datasets}
