% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_3_6.R
\docType{data}
\name{table_3_6}
\alias{table_3_6}
\title{Table 3.6 Productivity and Related Data, Business Sector 1960–2005}
\format{
A \link[tibble:tibble-package]{tibble} with 46 rows and
5 variables:
\describe{
\item{year}{}
\item{out_bus: }{Output per Hour of All Persons on Business Sector}
\item{out_nbus: }{Output per Hour of All Persons on Nonfarm Business Sector}
\item{rcom_bus: }{Real Compensation per Hour on Business Sector}
\item{rcom_nbus: }{Real Compensation per Hour on Nonfarm Business Sector}
}
}
\source{
Economic Report of the President, 2007, Table 49.

Note 1: Output refers to real gross domestic product in the sector.
Note 2: Wages and salaries of employees plus employers’ contributions for
social insurance and private beneﬁt plans.
Note 3: Hourly compensation divided by the consumer price index for all
urban consumers for recent quarters.
}
\usage{
table_3_6
}
\description{
Productivity and Related Data, Business Sector 1960–2005
(Index numbers, 1992 = 100; quarterly data seasonally adjusted)
}
\keyword{datasets}
