% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_6_9.R
\docType{data}
\name{table_6_9}
\alias{table_6_9}
\title{Table 6.9 Sample of 1,519 households drawn from the 1980–1982 British Family Expenditure Surveys}
\format{
A \link[tibble:tibble-package]{tibble} with 1519 rows and
16 variables:
\describe{
\item{wfood:}{budget share for food expenditure}
\item{wfuel:}{budget share for fuel expenditure}
\item{wcloth:}{budget share for clothing expenditure}
\item{wacl:}{budget share for alcohol expenditure}
\item{wtrans:}{budget share for transportation expenditure}
\item{wother:}{budget share for other expenditures}
\item{totexp:}{total household expenditure (rounded to the nearest 10 U.K. pounds sterling)}
\item{income:}{total net household income (rounded to the nearest 10 U.K. pounds sterling)}
\item{age:}{age of household head}
\item{nk:}{number of children}
\item{food:}{food expenditure}
\item{fuel:}{fuel expenditure}
\item{cloth:}{clothing expenditure}
\item{alc:}{alcohol expenditure}
\item{trans:}{transportation expenditure}
\item{other:}{other expenditures}
}
}
\source{
The data are from Richard Blundell and Krishna Pendakur, “Semiparametric Estimation and
Consumer Demand,” Journal of Applied Econometrics, vol. 13, no. 5, 1998, pp. 435–462.
Data reproduced from R. Carter Hill, William E. Grifﬁths, and George G. Judge,
Undergraduate Econometrics, 2d ed., John Wiley & Sons, New York, 2001.
}
\usage{
table_6_9
}
\description{
Sample of 1,519 households drawn from the 1980–1982 British Family Expenditure Surveys.
The data include only house-holds with one or two children living in Greater London.
The sample does not include self-employed or retired households.
}
\keyword{datasets}
