% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_13.R
\docType{data}
\name{table_7_13}
\alias{table_7_13}
\title{Table 7.13 Estimating Qualcomm stock prices}
\format{
A \link[tibble:tibble-package]{tibble} with 260 rows and
3 variables:
\itemize{
\item \strong{date}
\item \strong{time}
\item \strong{close}
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 226-227.
}
\usage{
table_7_13
}
\description{
Data on the weekly stock prices of Qualcomm, Inc., a digital wireless
telecommunications designer and manufacturer over the time period of 1995 to 2000
}
\keyword{datasets}
