% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_7.R
\docType{data}
\name{table_7_7}
\alias{table_7_7}
\title{Table 7.7 Wildcat Activity}
\format{
A \link[tibble:tibble-package]{tibble} with 31 rows and
6 variables:
\describe{
\item{y:}{number of wildcats drilled, thousands}
\item{x2:}{per barrel price, in constant $, 1972=100}
\item{x3:}{domestic output (millions of barrels per day)}
\item{x4:}{GNP, constant $ billions, 1972=100}
\item{x5:}{time: 1948 = 1, 1949 = 2, ... 1977 = 30, 1978 = 31}
\item{year}{}
}
}
\source{
Energy Information Administration, 1978 Report to Congress.
}
\usage{
table_7_7
}
\description{
Wildcat Activity
}
\keyword{datasets}
