% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_8.R
\docType{data}
\name{table_7_8}
\alias{table_7_8}
\title{Table 7.8 U.S. Defense Budget Outlays, 1962–1981}
\format{
A \link[tibble:tibble-package]{tibble} with 20 rows and
6 variables:
\itemize{
\item \strong{year}
\item \strong{y:} defense budget outlays
\item \strong{x2:} GNP
\item \strong{x3:} U.S. Military sales/assistance
\item \strong{x4:} aerospace industry sales
\item \strong{x5:} conflicts 100,000+
}
}
\source{
GUJARATI, Damodar N.; PORTER, Dawn C. Basic Econometrics, 5th ed.
Boston, Mass: McGraw-Hill, 2009, p. 221. These data were collected by
Albert Lucchino from various government publications.
}
\usage{
table_7_8
}
\description{
U.S. Defense Budget Outlays, 1962–1981
}
\keyword{datasets}
