% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_7_9.R
\docType{data}
\name{table_7_9}
\alias{table_7_9}
\title{Table 7.9 Demand for Chicken in the U.S., 1960–1982}
\format{
A \link[tibble:tibble-package]{tibble} with 23 rows and
7 variables:
\describe{
\item{year}{}
\item{y:}{per capita consumption of chickens, lb}
\item{x2:}{real disposable income per capita, $}
\item{x3:}{real retail price of chicken per lb, ¢}
\item{x4:}{real retail price of pork per lb, ¢}
\item{x5:}{real retail price of beef per lb, ¢}
\item{x6:}{composite real price of chicken substitutes per lb, ¢, which is a weighted average of the real retail prices per lb of pork and beef, the weights being the relative consumptions of beef and pork in total beef and pork consumption}
}
}
\source{
\url{http://lib.stat.cmu.edu/DASL/Dataﬁles/SmokingandCancer.html}
}
\usage{
table_7_9
}
\description{
Demand for Chicken in the U.S., 1960–1982
Note: The real prices were obtained by dividing the nominal prices
by the Consumer Price Index for food
}
\keyword{datasets}
